Defer full English catalog off homepage bootstrap
