+++ client/src/lib/i18n.full.tsx
@@ -1,0 +1,1763 @@
+import { createContext, useContext, useState, useCallback, useEffect, ReactNode } from 'react';
+
+export type Language = 'en' | 'th';
+type TranslationCatalog = Record<string, string>;
+
+interface I18nContextType {
+  lang: Language;
+  setLang: (lang: Language) => void;
+  t: (key: string, params?: Record<string, string | number>) => string;
+}
+
+const I18nContext = createContext<I18nContextType | null>(null);
+
+export function detectLanguage(): Language {
+  if (typeof window === 'undefined') return 'en';
+
+  const saved = localStorage.getItem('thaichess-lang');
+  if (saved === 'th' || saved === 'en') return saved;
+
+  const browserLang = navigator.language || (navigator as any).userLanguage || '';
+  if (browserLang.startsWith('th')) return 'th';
+
+  return 'en';
+}
+
+async function loadThaiTranslations(): Promise<TranslationCatalog> {
+  const module = await import('./i18n.th');
+  return module.TH_TRANSLATIONS;
+}
+
+function applyParams(text: string, params?: Record<string, string | number>): string {
+  if (params) {
+    for (const [k, v] of Object.entries(params)) {
+      text = text.replace(new RegExp(`\\{${k}\\}`, 'g'), String(v));
+    }
+  }
+  return text;
+}
+
+export const EN_TRANSLATIONS: Record<string, string> = {
+  // Common
+  'app.name': 'ThaiChess',
+  'app.tagline': 'The Ancient Art of Chess',
+  'nav.play': 'Play',
+  'nav.watch': 'Watch',
+  'nav.lessons': 'Lessons',
+  'nav.puzzles': 'Puzzles',
+  'nav.games': 'Games',
+  'nav.about': 'About',
+  'common.back_home': 'Back to Home',
+  'common.white': 'White',
+  'common.black': 'Black',
+  'common.draw': 'Draw',
+  'common.you': 'You',
+  'common.guest': 'Guest',
+  'common.anonymous': 'Anonymous',
+  'common.cancel': 'Cancel',
+  'common.send': 'Send',
+  'common.sending': 'Sending...',
+  'common.close': 'Close',
+  'common.retry': 'Retry',
+  'common.new_game': 'New Game',
+  'common.loading': 'Loading...',
+  'footer.tagline': 'ThaiChess — Free & Open Source',
+  'footer.inspired': 'Inspired by',
+  'footer.community': 'Community',
+  'footer.github': 'GitHub',
+  'footer.star_github': 'Star on GitHub',
+
+  // Appearance
+  'appearance.title': 'Board & Pieces',
+  'appearance.board_and_pieces': 'Board & Pieces',
+  'appearance.subtitle': 'Choose a board theme and a readable piece color theme once, then use them everywhere: online, bot, local, and puzzles.',
+  'appearance.open': 'Board & Pieces',
+  'appearance.open_short': 'Theme',
+  'appearance.boards_tab': 'Boards',
+  'appearance.colors_tab': 'Piece Colors',
+  'appearance.live_preview': 'Live Preview',
+  'appearance.preview_title': 'Preview',
+  'appearance.preview_subtitle': 'Hover to preview, click to apply instantly. Your choice is saved automatically on this device.',
+  'appearance.preview_board': 'Board theme',
+  'appearance.preview_piece_theme': 'Piece color theme',
+  'appearance.core_shape_label': 'Core shape',
+  'appearance.core_shape_desc': 'is the single supported gameplay silhouette for fast recognition and consistent play.',
+  'appearance.saved_note': 'Saved automatically on this device',
+  'appearance.category_classic': 'Classic',
+  'appearance.category_soft': 'Soft',
+  'appearance.category_dark': 'Dark Mode',
+  'appearance.category_elegant': 'Elegant',
+  'appearance.readability_checked': 'Contrast checked',
+  'appearance.grid_contrast': 'Grid contrast',
+  'appearance.piece_contrast': 'Weakest piece contrast',
+  'appearance.single_surface_label': 'Single-surface Makruk',
+  'appearance.single_surface_desc': 'Themes now change board tone, texture feel, and grid only.',
+  'appearance.comparison_title': 'Old vs Makruk',
+  'appearance.comparison_desc': 'Themes no longer switch square colors. They only change the board material tone and grid.',
+  'appearance.comparison_before': 'Before',
+  'appearance.comparison_after': 'Makruk now',
+
+  // Home Page
+  'home.hero_title': 'Play Makruk Instantly',
+  'home.hero_desc': 'No signup required. Start a ThaiChess game in seconds, play with friends, or practice against the bot.',
+  'home.quick_play': 'Play Now',
+  'home.quick_play_desc': 'Get paired instantly for casual or rated Makruk.',
+  'home.find_opponent': 'Find Opponent',
+  'home.no_signup': 'No signup required',
+  'home.free_to_play': 'Free to play',
+  'home.games_played': '{count} games played',
+  'home.play_friend': 'Play a Friend',
+  'home.play_friend_desc': 'Share a link',
+  'home.play_bot': 'Play vs Bot',
+  'home.play_bot_desc': '10 themed personas',
+  'home.play_bot_long_desc': 'Challenge a roster of distinct Makruk rivals, each with a named identity, rating, and signature style.',
+  'home.puzzles': 'Puzzles',
+  'home.puzzles_desc': 'Tactical training',
+  'home.puzzles_long_desc': 'Sharpen your tactical skills with curated puzzles from real games.',
+  'home.lessons': 'Lessons',
+  'home.lessons_desc': 'Structured Makruk course',
+  'home.create_private': 'Create a Private Game',
+  'home.private_desc': 'Choose a time control, pick a color, and share the game link.',
+  'home.time_control': 'Time Control',
+  'home.choose_color': 'Choose Color',
+  'home.color_random': 'Random',
+  'home.color_white': 'White',
+  'home.color_black': 'Black',
+  'home.play_with_friend': 'Play with a Friend',
+  'home.creating': 'Creating...',
+  'home.or': 'or',
+  'home.play_local': 'Play Locally (Same Screen)',
+  'home.play_local_desc': 'Pass the board back and forth on one screen.',
+  'home.watch_live': 'Watch Live Games',
+  'home.watch_live_desc': 'See active public Makruk games and jump straight into spectator mode.',
+  'home.live_now_title': 'Live Now',
+  'home.live_now_desc': 'Active public games update in real time. Open any board in read-only spectator mode.',
+  'home.view_all_live': 'View All Live Games',
+  'home.no_live_games': 'No live games right now',
+  'home.no_live_games_desc': 'Check back soon for active public matches to watch.',
+  'home.join_prompt': 'Have a game code?',
+  'home.join_link': 'Join a game',
+  'home.join_title': 'Join a Game',
+  'home.join_desc': 'Open a shared game instantly with a code from a friend.',
+  'home.join_placeholder': 'Enter game code...',
+  'home.join': 'Join',
+  'home.training_start': 'Start puzzle track',
+  'home.training_continue': 'Continue training',
+  'home.training_progress': '{done}/{total} puzzles completed',
+  'home.training_focus': 'Strongest theme so far: {theme}',
+  'home.training_resume': 'Last played: {title}',
+  'home.training_recent': 'Latest solve: {title}',
+  'home.streak_start': 'Start streak',
+  'home.streak_continue': 'Jump back in',
+  'home.streak_title': 'Puzzle streak',
+  'home.streak_progress': '{done}/{total} lessons solved',
+  'home.streak_focus': 'Best lesson theme so far: {theme}',
+  'home.streak_resume': 'Last lesson played: {title}',
+  'home.streak_recent': 'Latest lesson solved: {title}',
+  'home.learn_eyebrow': 'Learn',
+  'home.learn_title': 'Start With the Pages That Actually Help',
+  'home.learn_desc': 'If you are new here, these three pages take you from basic context to your first real Makruk session.',
+  'home.learn_card.what_is_title': 'What Is Makruk?',
+  'home.learn_card.what_is_desc': 'Get the big picture first and see why Thai chess feels different from western chess.',
+  'home.learn_card.how_to_title': 'How to Play Makruk',
+  'home.learn_card.how_to_desc': 'Learn piece movement, promotion, and the counting rule without the usual confusion.',
+  'home.learn_card.play_online_title': 'Play Makruk Online',
+  'home.learn_card.play_online_desc': 'See whether bot games, puzzles, or live play make the best first step for you.',
+  'lessons.course.eyebrow': 'Structured lessons',
+  'lessons.course.title': 'Learn Makruk like a coached course',
+  'lessons.course.desc': 'Work through beginner, intermediate, and advanced modules in order. Every lesson explains one idea, shows it on the board, guides you step by step, and then asks you to practice it right away.',
+  'lessons.course.stats.progress': 'Progress',
+  'lessons.course.stats.completed': 'Lessons completed',
+  'lessons.course.stats.shape': 'Course shape',
+  'lessons.course.stats.shape_desc': 'Modules from basics to planning',
+  'lessons.course.stats.format': 'Format',
+  'lessons.course.stats.format_value': 'Coach-led',
+  'lessons.course.stats.format_desc': 'Explain, demonstrate, practice, reinforce',
+  'lessons.course.progress_label': 'Course progress',
+  'lessons.course.progress_value': '{percent}% complete',
+  'lessons.course.next_lesson': 'Next lesson',
+  'lessons.course.minutes': '{minutes} min',
+  'lessons.course.start_lesson': 'Start lesson',
+  'lessons.course.complete_title': 'Course complete',
+  'lessons.course.complete_desc': 'You have finished the current Makruk course. Replay any lesson to refresh the ideas, then head into puzzles to keep the patterns sharp.',
+  'lessons.course.go_to_puzzles': 'Go to puzzles',
+  'lessons.course.module_progress': '{done}/{total} complete',
+  'lessons.course.lesson_label': 'Lesson {order}',
+  'lessons.course.status.completed': 'Completed',
+  'lessons.course.status.resume': 'Resume',
+  'lessons.course.status.unlocked': 'Unlocked',
+  'lessons.course.status.locked': 'Locked',
+  'lessons.player.not_found': 'Lesson not found',
+  'lessons.player.back_to_course': 'Back to course',
+  'lessons.player.default_correct': 'Correct.',
+  'lessons.player.default_wrong': 'That move does not match the lesson target.',
+  'lessons.player.practice_task': 'Practice task',
+  'lessons.player.guided_step': 'Guided step',
+  'lessons.player.mini_practice': 'Mini practice',
+  'lessons.player.status.completed': 'Completed',
+  'lessons.player.status.guided': 'Guided',
+  'lessons.player.status.practice': 'Practice',
+  'lessons.player.header_subtitle': 'Lesson · {title}',
+  'lessons.player.course_path': 'Course path',
+  'lessons.player.level_and_order': '{level} · Lesson {order}',
+  'lessons.player.counting_aware': 'Counting-aware lesson',
+  'lessons.player.no_counting_dependency': 'No counting dependency',
+  'lessons.player.lesson_flow': 'Lesson flow',
+  'lessons.player.step_of': 'Step {current} of {total}',
+  'lessons.player.task_of': 'Task {current} of {total}',
+  'lessons.player.concept_explanation': 'Concept explanation',
+  'lessons.player.rule_impact': 'Rule impact',
+  'lessons.player.counting_changes_evaluation': 'Counting changes the evaluation',
+  'lessons.player.counting_not_relevant': 'Counting does not change this lesson',
+  'lessons.player.why_this_move_works': 'Why this move works',
+  'lessons.player.problem': 'Problem',
+  'lessons.player.what_it_fixes': 'What it fixes',
+  'lessons.player.new_threat': 'New threat',
+  'lessons.player.after_the_move': 'After the move',
+  'lessons.player.compare_moves': 'Compare moves',
+  'lessons.player.best_move': 'Best move',
+  'lessons.player.tempting_move': 'Tempting move',
+  'lessons.player.step_badge': 'Step {number}',
+  'lessons.player.task_badge': 'Task {number}',
+  'lessons.player.start_mini_practice': 'Start mini practice',
+  'lessons.player.next_step': 'Next step',
+  'lessons.player.review_summary': 'Review summary',
+  'lessons.player.next_task': 'Next task',
+  'lessons.player.next_lesson': 'Next lesson',
+  'lessons.player.summary': 'Summary',
+  'lessons.player.puzzles_reinforce': 'Puzzles to reinforce this lesson',
+  'lessons.player.puzzles_reinforce_desc': 'After the lesson, jump into related puzzle themes so the idea moves from explanation to pattern recognition.',
+  'lessons.player.more_puzzles_coming': 'More puzzle links will appear here as the tactic catalog expands around this concept.',
+  'lessons.player.previous_lesson': 'Previous lesson',
+
+  // Time presets
+  'time.bullet': 'Bullet',
+  'time.blitz': 'Blitz',
+  'time.rapid': 'Rapid',
+  'time.classical': 'Classical',
+
+  // Bot Game
+  'bot.title': 'Play vs Bot',
+  'bot.setup_title': 'Play vs Computer',
+  'bot.setup_desc': 'Choose a bot persona, review its style, level, and estimated strength, pick your side, then begin. Random side is decided only when the game begins.',
+  'bot.selected_bot': 'Selected Bot',
+  'bot.roster': 'Bot roster',
+  'bot.featured_opponent': 'Featured opponent',
+  'bot.opening_preference': 'Opening preference',
+  'bot.signature_style': 'Signature style',
+  'bot.tactical_bias': 'Tactical bias',
+  'bot.strategic_weakness': 'Strategic weakness',
+  'bot.dialogue_preview': 'Dialogue preview',
+  'bot.difficulty': 'Difficulty',
+  'bot.level_label': 'Level',
+  'bot.level_short': 'Level {level}',
+  'bot.estimated_elo_label': 'Estimated ELO',
+  'bot.estimated_elo_range': 'Estimated {range} ELO',
+  'bot.estimated_elo_note': 'Estimated strength based on play behavior, not an official rating.',
+  'bot.easy': 'Easy',
+  'bot.easy_desc': 'Random moves with basic captures',
+  'bot.medium': 'Medium',
+  'bot.medium_desc': 'Thinks 2 moves ahead',
+  'bot.hard': 'Hard',
+  'bot.hard_desc': 'Thinks 4 moves ahead',
+  'bot.play_as': 'Play as',
+  'bot.random': 'Random',
+  'bot.random_assigned': 'Assigned on start',
+  'bot.side_locked': 'Locked before game start',
+  'bot.start': 'Start Game',
+  'bot.thinking': 'thinking...',
+  'bot.chat_thinking': 'A quiet thought',
+  'bot.your_turn': 'Your turn',
+  'bot.bot_thinking': 'Bot is thinking...',
+  'bot.you_won': 'You Won!',
+  'bot.bot_wins': 'Bot Wins',
+  'bot.resign': 'Resign',
+  'bot.resign_confirm': 'Are you sure you want to resign?',
+  'bot.vs_bot': 'vs Bot',
+
+  // Puzzles
+  'puzzle.title': 'Puzzle Streak',
+  'puzzle.desc': 'Play one tactical puzzle after another with hidden adaptive difficulty and instant scoring.',
+  'puzzle.completed': '{done}/{total} completed',
+  'puzzle.completed_summary': 'Overall progress',
+  'puzzle.progress_hint': 'Solved puzzles stay checked on this device.',
+  'puzzle.all': 'All',
+  'puzzle.all_lessons': 'All Lessons',
+  'puzzle.beginner': 'Beginner',
+  'puzzle.intermediate': 'Intermediate',
+  'puzzle.advanced': 'Advanced',
+  'puzzle.filter_label': 'Current track',
+  'puzzle.filter_all_desc': 'Mix every shipped lesson in one queue.',
+  'puzzle.filter_beginner_desc': 'Short tactical wins and basic mates.',
+  'puzzle.filter_intermediate_desc': 'Two-step ideas that punish weak defense.',
+  'puzzle.filter_advanced_desc': 'Tougher calculation once the catalog grows.',
+  'puzzle.theme_drill_title': 'Theme drills',
+  'puzzle.theme_drill_desc': 'Narrow the queue to one tactical idea when you want repetition.',
+  'puzzle.theme_drill_count': '{count} themes in this track',
+  'puzzle.theme_all': 'All ideas',
+  'puzzle.track_progress': '{track} progress',
+  'puzzle.track_completed': '{done}/{total} in this track',
+  'puzzle.remaining_label': 'Still fresh',
+  'puzzle.remaining_desc': 'Unsolved puzzles are listed first.',
+  'puzzle.focus_label': 'Theme focus',
+  'puzzle.focus_desc': 'Use this to pick what idea you want to drill next.',
+  'puzzle.focus_empty': 'No themes yet',
+  'puzzle.next_up': 'Next lesson',
+  'puzzle.next_up_fresh': 'Start with the first unsolved puzzle in this track.',
+  'puzzle.next_up_review': 'You cleared this track. Replay a solved puzzle to stay sharp.',
+  'puzzle.start_here': 'Start here',
+  'puzzle.practice_title': 'Practice queue',
+  'puzzle.practice_desc': 'Fresh puzzles come first, solved ones stay available for review.',
+  'puzzle.new_badge': 'Fresh',
+  'puzzle.solved_badge': 'Solved',
+  'puzzle.empty_title': 'No puzzles in this track yet',
+  'puzzle.empty_desc': 'Switch difficulty or come back after more lessons are reviewed.',
+  'puzzle.empty_theme_title': 'No {theme} drills here yet',
+  'puzzle.empty_theme_desc': 'Try another idea or switch out of {track} to find more {theme} puzzles.',
+  'puzzle.to_move': '{color} to move',
+  'puzzle.find_best': 'Find the best move for {color}!',
+  'puzzle.step': 'Step {current} of {total}',
+  'puzzle.correct': 'Correct!',
+  'puzzle.solved_hint': 'Solved with hint',
+  'puzzle.solved_clean': 'Solved without hints!',
+  'puzzle.wrong': 'Not quite!',
+  'puzzle.wrong_desc': "That wasn't the best move. Try again!",
+  'puzzle.hint': 'Hint',
+  'puzzle.next': 'Next Puzzle',
+  'puzzle.previous': 'Previous',
+  'puzzle.all_puzzles': 'All Puzzles',
+  'puzzle.back_to_lessons': 'Back to Lessons',
+  'puzzle.not_found': 'Puzzle not found',
+  'puzzle.back': 'Back to Puzzles',
+  'puzzle.lesson': 'Lesson',
+  'puzzle.lessons_nav': 'Lessons',
+  'puzzle.lessons_eyebrow': 'Lesson Mode',
+  'puzzle.lessons_title': 'Lessons',
+  'puzzle.lessons_desc': 'Learn Makruk through a progressive course, then reinforce each concept with puzzles.',
+  'puzzle.lessons_tracks_title': 'Learning tracks',
+  'puzzle.lessons_tracks_desc': 'Beginner, Intermediate, and Advanced stay here as deliberate study paths.',
+  'puzzle.play_streak': 'Play Streak',
+  'puzzle.streak_nav': 'Puzzle Streak',
+  'puzzle.streak_eyebrow': 'Primary Mode',
+  'puzzle.streak_title': 'Puzzle Streak',
+  'puzzle.streak_desc': 'Solve continuously, build your streak, and let the system quietly raise or lower the challenge based on your play.',
+  'puzzle.streak_prompt_title': 'Keep the streak alive',
+  'puzzle.streak_prompt_desc': 'Correct moves score immediately and the next puzzle loads right away.',
+  'puzzle.streak_score_label': 'Score',
+  'puzzle.streak_label': 'Streak',
+  'puzzle.streak_session_label': 'Solved',
+  'puzzle.streak_checkpoint_label': 'Checkpoint',
+  'puzzle.streak_checkpoint_progress': '{current}/{total} to the next pulse',
+  'puzzle.streak_flow_label': 'Flow',
+  'puzzle.streak_flow_desc': 'Adaptive pacing',
+  'puzzle.streak_puzzle_label': 'Puzzle {number}',
+  'puzzle.streak_points': '+{points} points',
+  'puzzle.streak_focus_title': 'Current puzzle',
+  'puzzle.streak_focus_keep': 'Breathe, reset, and go again.',
+  'puzzle.streak_focus_desc': 'Difficulty stays hidden here so the rhythm feels fast and frictionless.',
+  'puzzle.streak_session_title': 'Session',
+  'puzzle.streak_best_label': 'Best streak',
+  'puzzle.streak_lessons_progress_label': 'Lessons',
+  'puzzle.visit_lessons': 'Open Lessons',
+  'puzzle.streak_end_eyebrow': 'Session Ended',
+  'puzzle.streak_end_title': 'Streak ended',
+  'puzzle.streak_end_desc': 'One mistake breaks the run, but your lesson progress stays saved.',
+  'puzzle.streak_ended_at': 'Streak ended at {streak}',
+  'puzzle.streak_end_summary': 'Final streak {streak} · session score {score}',
+  'puzzle.retry_puzzle': 'Retry puzzle',
+  'puzzle.streak_try_again': 'Start a new streak',
+  'puzzle.streak_milestone_improving': 'You’re improving. Keep the flow going.',
+  'puzzle.streak_milestone_harder': 'Now facing harder puzzles. Stay sharp.',
+  'puzzle.source': 'Source',
+  'puzzle.source_real_game_ply': 'Real game · ply {ply}',
+  'puzzle.source_seed_game_ply': 'Seed game · ply {ply}',
+  'puzzle.source_starter_pack': 'Starter pack',
+  'puzzle.source_curated_tactic': 'Curated tactic',
+  'puzzle.source_tactical_motif': 'Tactical motif',
+  'puzzle.source_review_batch': 'Review batch',
+  'puzzle.rating_label': 'Adaptive target',
+  'puzzle.rating_short': 'Rating {score}',
+  'puzzle.success_rate': 'Success rate',
+  'puzzle.attempts_label': 'Attempts',
+  'puzzle.activity_title': 'Session activity',
+  'puzzle.activity_status_label': 'Status',
+  'puzzle.activity_status_new': 'New',
+  'puzzle.activity_status_in_progress': 'In progress',
+  'puzzle.activity_status_solved': 'Solved',
+  'puzzle.activity_last_played': 'Last played {date}',
+  'puzzle.activity_completed_on': 'Solved {date}',
+  'puzzle.related_theme_title': 'More in this theme',
+  'puzzle.related_theme_desc': 'Keep drilling {theme} with these follow-ups.',
+  'puzzle.related_theme_empty': 'No other {theme} puzzles yet.',
+
+  // Puzzle themes
+  'theme.BasicCheckmate': 'Basic Checkmate',
+  'theme.BackRank': 'Back Rank',
+  'theme.Checkmate': 'Checkmate',
+  'theme.Clearance': 'Clearance',
+  'theme.Decoy': 'Decoy',
+  'theme.Defense': 'Defense',
+  'theme.BestDefense': 'Best Defense',
+  'theme.Deflection': 'Deflection',
+  'theme.CountDefense': 'Count Defense',
+  'theme.CountingDraw': 'Counting Draw',
+  'theme.DoubleAttack': 'Double Attack',
+  'theme.DoubleCheck': 'Double Check',
+  'theme.ExchangeSacrifice': 'Exchange Sacrifice',
+  'theme.HangingPiece': 'Hanging Piece',
+  'theme.Interference': 'Interference',
+  'theme.MateIn1': 'Mate in 1',
+  'theme.MateIn2': 'Mate in 2',
+  'theme.MateIn3': 'Mate in 3+',
+  'theme.MatingNet': 'Mating Net',
+  'theme.Overloading': 'Overloading',
+  'theme.PerpetualCheck': 'Perpetual Check',
+  'theme.Fork': 'Fork',
+  'theme.Pin': 'Pin',
+  'theme.Promotion': 'Promotion',
+  'theme.Discovery': 'Discovery',
+  'theme.RemovalOfDefender': 'Removal of the Defender',
+  'theme.OnlyMove': 'Only Move',
+  'theme.Sacrifice': 'Sacrifice',
+  'theme.SaveTheDraw': 'Save the Draw',
+  'theme.Simplification': 'Simplification',
+  'theme.SmotheredMate': 'Smothered Mate',
+  'theme.Skewer': 'Skewer',
+  'theme.Stalemate': 'Stalemate',
+  'theme.SupportMate': 'Support Mate',
+  'theme.Tactic': 'Tactic',
+  'theme.TacticalWin': 'Tactical Win',
+  'theme.TrappedPiece': 'Trapped Piece',
+  'theme.VulnerableKing': 'Vulnerable King',
+  'theme.WinBeforeCountExpires': 'Win Before Count Expires',
+  'theme.Windmill': 'Windmill',
+  'theme.XRay': 'X-Ray Attack',
+  'theme.Zwischenzug': 'Zwischenzug',
+  'theme.Endgame': 'Endgame',
+  'theme.Desperado': 'Desperado',
+
+  // Quick Play / Matchmaking
+  'quick.title': 'Quick Play',
+  'quick.desc': 'Find an opponent instantly. No link sharing needed!',
+  'quick.searching': 'Finding opponent...',
+  'quick.search_time': 'Searching for {time}',
+  'quick.queue': '{count} player(s) in queue',
+  'quick.find': 'Find Opponent',
+  'quick.rated_available': 'Rated Available',
+  'quick.rated_unavailable': 'Rated Disabled',
+  'quick.casual_only': 'Casual Only',
+  'quick.load_failed': 'Unable to start quick play right now.',
+  'quick.rated_signed_in': 'Rated if your opponent is also signed in.',
+  'quick.rated_restricted': 'This account can still quick-play casually, but rated pairings are disabled.',
+  'quick.rated_sign_in': 'Sign in to unlock rated games.',
+
+  // Game Page
+  'game.waiting_title': 'Waiting for opponent',
+  'game.waiting_desc': 'Share this link with a friend to start playing',
+  'game.copy': 'Copy',
+  'game.copied': 'Copied!',
+  'game.playing_as': 'You are playing as {color}',
+  'game.rated': 'Rated',
+  'game.casual': 'Casual',
+  'game.rating_change': 'Rating',
+  'game.connecting': 'Connecting to game...',
+  'game.load_failed': 'Unable to load this game right now.',
+  'game.error': 'Error',
+  'game.opponent': 'Opponent',
+  'game.online': 'Online',
+  'game.offline': 'Offline',
+  'game.active_now': 'Active',
+  'game.idle': 'Idle',
+  'game.away': 'Away',
+  'game.ping_short': 'Ping',
+  'game.ping_unknown': 'Ping unavailable',
+  'game.ping_value': '{ms} ms latency',
+  'game.to_move': 'To move',
+  'game.your_turn': 'Your turn',
+  'game.opponent_turn': "Opponent's turn",
+  'game.you_won': 'You won!',
+  'game.you_lost': 'You lost',
+  'game.draw_offer_received': 'Your opponent offers a draw',
+  'game.accept': 'Accept',
+  'game.decline': 'Decline',
+  'game.offer_draw': '½ Draw',
+  'game.resign': '⚐ Resign',
+  'game.resign_confirm': 'Are you sure you want to resign?',
+  'game.share': 'Share',
+  'game.open_spectator': 'Open Spectator View',
+  'game.piece_guide': '📖 Piece Guide',
+  'game.opponent_dc': 'Opponent disconnected. Waiting for reconnection...',
+  'game.game_label': 'Game:',
+  'game.counting_title': 'Counting Rule',
+  'game.counting_board_honor': 'Board honor',
+  'game.counting_pieces_honor': 'Pieces honor',
+  'game.counting_status': '{type}: {color} is counting {current}/{limit}',
+  'game.counting_final': '{type}: final attacking move before draw',
+  'game.counting_available': '{type}: {color} may start counting',
+  'game.counting_start': 'Start Count',
+  'game.counting_stop': 'Stop Count',
+  'game.premove_set': 'Pre-move set',
+  'game.check_status': 'Check on board',
+  'game.nav_hint': 'Use arrow keys to navigate moves',
+  'game.select_piece_style': 'Select piece style',
+  'game.piece_style': 'Pieces',
+  'game.piece_style_current': 'Current',
+  'game.piece_style_western': 'Western',
+  'game.piece_style_makruk': 'Makruk',
+  'game.match_status': 'Match Status',
+  'game.playing_as_label': 'Playing as',
+  'game.review_mode': 'Board View',
+  'game.reviewing_history': 'Reviewing history',
+  'game.live_position': 'Live position',
+  'game.return_to_live': 'Back to current move',
+  'game.live_match_title': 'Live Makruk Match',
+  'game.live_match_desc': 'The board comes first. Clocks, review, and controls stay close without competing with the game.',
+  'game.spectator_mode': 'Spectator Mode',
+  'game.watching_live_game': 'Watching Live Game',
+  'game.spectator_desc': 'Read-only live view with board updates, clocks, and the move list.',
+  'game.live_now': 'Live',
+  'game.final_position': 'Final Position',
+  'game.read_only': 'Read-only',
+  'game.flip_board': 'Flip Board',
+  'game.white_perspective': 'White View',
+  'game.black_perspective': 'Black View',
+  'game.live_position_locked': 'The board stays on the current live position while the game is running.',
+  'game.spectator_turn': '{color} to move',
+  'game.spectator_controls_hidden': 'Player controls are hidden in spectator mode.',
+  'game.post_game_title': 'Review the Finished Game',
+  'game.post_game_desc': 'Step through the moves, study the ending, or jump straight into analysis.',
+  'game.side_title': 'Match Panel',
+  'game.you_label': 'You',
+  'game.reviewing_position': 'Finished game',
+
+  // Game Over Modal
+  'gameover.you_win': 'You Win!',
+  'gameover.you_lost': 'You Lost',
+  'gameover.draw': 'Draw',
+  'gameover.by_checkmate': 'by checkmate',
+  'gameover.by_resign': 'by resignation',
+  'gameover.by_timeout': 'on time',
+  'gameover.by_stalemate': 'by stalemate',
+  'gameover.by_agreement': 'by mutual agreement',
+  'gameover.by_material': 'by insufficient material',
+  'gameover.by_counting': 'by counting rule',
+  'gameover.by_unknown': 'game ended',
+  'gameover.rematch': 'Rematch',
+  'gameover.rematch_accept': 'Accept Rematch',
+  'gameover.rematch_sent': 'Rematch Sent',
+  'gameover.rematch_waiting': 'Rematch request sent. Waiting for your opponent.',
+  'gameover.rematch_incoming': 'Your opponent wants a rematch.',
+  'gameover.is_victorious': 'is victorious',
+
+  // Local Game
+  'local.title': 'Local Game',
+  'local.view_as': 'View as:',
+  'local.turn': "{color}'s turn",
+  'local.wins': '{color} wins!',
+  'local.play_online': 'Play Online',
+
+  // Move History
+  'moves.title': 'Moves',
+  'moves.empty': 'No moves yet',
+  'moves.first_position': 'First position',
+  'moves.previous_move': 'Previous move',
+  'moves.next_move': 'Next move',
+  'moves.last_move': 'Last move',
+
+  // Piece Guide
+  'guide.title': 'Piece Guide',
+  'guide.king': 'Khun (King)',
+  'guide.queen': 'Met (Queen)',
+  'guide.bishop': 'Khon (Bishop)',
+  'guide.rook': 'Rua (Rook)',
+  'guide.knight': 'Ma (Knight)',
+  'guide.pawn': 'Bia (Pawn)',
+  'guide.promoted': 'Bia Ngai (Promoted)',
+  'guide.king_move': '1 square in any direction',
+  'guide.queen_move': '1 square diagonally only',
+  'guide.bishop_move': '1 square diagonally or 1 forward',
+  'guide.rook_move': 'Any distance horizontally or vertically',
+  'guide.knight_move': 'L-shape: 2+1 squares',
+  'guide.pawn_move': '1 forward; captures diagonally',
+  'guide.promoted_move': 'Same as Met (1 diagonal)',
+
+  // Connection
+  'conn.connected': 'Connected',
+  'conn.reconnecting': 'Reconnecting...',
+  'conn.disconnected': 'Disconnected — trying to reconnect...',
+
+  // About Page
+  'about.eyebrow': 'Learn Makruk',
+  'about.title': 'About ThaiChess',
+  'about.hero_desc': 'Makruk is Thailand’s traditional chess: slower to explode, harder to simplify, and full of unfamiliar piece dynamics. ThaiChess exists to make that experience easy to learn and easy to play online.',
+  'about.cta_play': 'Play Now',
+  'about.cta_puzzles': 'Try Puzzles',
+  'about.quick_facts': 'Quick Facts',
+  'about.fact_heritage_label': 'Origin',
+  'about.fact_heritage_value': 'Traditional Thai chess',
+  'about.fact_style_label': 'Feel',
+  'about.fact_style_value': 'Positional, tactical, deliberate',
+  'about.intro_title': 'Why This Site Exists',
+  'about.intro1': 'ThaiChess is a <strong>free, open-source</strong> platform for playing ThaiChess (Makruk) with anyone in the world. No registration, no ads, no paywall — just pure Thai chess.',
+  'about.intro2': 'Our mission is to make <strong>ThaiChess famous worldwide</strong>. Thai chess is one of the oldest board games in the world, closer to the original Indian chess (Chaturanga) than modern Western chess. It deserves a world-class online platform.',
+  'about.intro3': 'Inspired by <a href="https://lichess.org" target="_blank" rel="noopener" class="text-primary hover:text-primary-light underline">Lichess</a> and <a href="https://chess.com" target="_blank" rel="noopener" class="text-primary hover:text-primary-light underline">Chess.com</a>, we believe great gaming platforms should be accessible to everyone.',
+  'about.stats_title': 'Platform Stats',
+  'about.games_played': 'Games Played',
+  'about.total_moves': 'Total Moves',
+  'about.white_wins': 'White Wins',
+  'about.black_wins': 'Black Wins',
+  'about.what_title': 'What is ThaiChess?',
+  'about.what1': '<strong>ThaiChess (หมากรุก)</strong> is the traditional chess of Thailand. It evolved from the ancient Indian game Chaturanga and has been played in Thailand for centuries.',
+  'about.what2': 'Unlike Western chess, ThaiChess retains many features of the original game — the queen (Met) is weak, the bishop (Khon) moves differently, and pawns start on the third rank. This creates a game that is deeply strategic and uniquely beautiful.',
+  'about.what3': "ThaiChess is widely played in Thailand and Cambodia (where it's called Ouk Chatrang). It is recognized by the World Chess Federation and has international tournaments.",
+  'about.what4': 'We believe ThaiChess deserves the same global recognition as chess. Help us spread the word!',
+  'about.diff_eyebrow': 'What Changes',
+  'about.diff_title': 'How Makruk feels different',
+  'about.diff1': 'The Met is much weaker than a western queen, so attacks grow more slowly and positioning matters more.',
+  'about.diff2': 'Khon movement changes how diagonals work, which creates unfamiliar tactical patterns even for strong chess players.',
+  'about.diff3': 'Pawns begin on the third rank and promote differently, so openings and endgames develop on their own terms.',
+  'about.diff4': 'The counting rule gives certain endgames a very different rhythm from western chess and rewards precision.',
+  'about.pieces_eyebrow': 'Piece Basics',
+  'about.pieces_title': 'Learn the movement patterns fast',
+  'about.learn_by_playing': 'Learn by playing a quick game',
+  'about.opensource_title': 'Open Source',
+  'about.opensource_desc': 'ThaiChess is 100% open source under the MIT license. The entire codebase is available on GitHub.',
+  'about.footer': 'ThaiChess — Free & Open Source — Made with love for Thai Chess',
+
+  // Games Page
+  'games.title': 'Recent Games',
+  'games.count': '{count} game(s) played',
+  'games.empty': 'No games yet',
+  'games.empty_desc': 'Be the first to play!',
+  'games.start': 'Start Playing',
+  'games.featured_title': 'Worth reviewing first',
+  'games.featured_desc': 'Longer or more complete games stay up top so you can jump into better analysis faster.',
+  'games.no_featured_title': 'Mostly quick endings on this page',
+  'games.no_featured_desc': 'These recent results are mostly agreements, instant resignations, or early timeouts.',
+  'games.low_signal_title': 'Quick endings and low-signal games',
+  'games.low_signal_desc': 'These games still count, but they usually teach less because they ended almost immediately.',
+  'games.low_signal_badge': 'Quick end',
+  'games.col_game': 'Game',
+  'games.col_time': 'Time',
+  'games.col_result': 'Result',
+  'games.col_moves': 'Moves',
+  'games.col_when': 'When',
+  'games.prev': 'Prev',
+  'games.next': 'Next',
+  'games.page': 'Page {current} of {total}',
+  'games.view_leaderboard': 'Leaderboard',
+  'games.filter_all': 'All',
+  'games.filter_rated': 'Rated',
+  'games.filter_casual': 'Casual',
+  'games.filter_bot': 'Bot',
+  'games.bot_badge': 'Bot',
+  'games.bot_games': 'Games vs bot',
+  'games.bot_win_rate': 'Win rate vs bot',
+  'games.bot_highest_level': 'Highest level beaten',
+  'games.reason_checkmate': 'Checkmate',
+  'games.reason_resignation': 'Resignation',
+  'games.reason_timeout': 'Timeout',
+  'games.reason_stalemate': 'Stalemate',
+  'games.reason_agreement': 'Agreement',
+  'games.reason_material': 'Insufficient material',
+  'games.reason_counting': 'Counting rule',
+  'games.reason_draw': 'Draw',
+  'games.reason_unknown': 'Finished',
+  'live.title': 'Watch Live Games',
+  'live.desc': 'Discover active public Makruk games and follow them move by move in spectator mode.',
+  'live.live_now': 'Live Now',
+  'live.live_count': '{count} live game(s)',
+  'live.recently_finished': 'Recently Finished',
+  'live.recently_finished_desc': 'These public games just ended and are still available for review.',
+  'live.empty_title': 'No live games at the moment',
+  'live.empty_desc': 'When public quick-play games start, they will appear here automatically.',
+  'live.card_live': 'Live',
+  'live.card_finished': 'Finished',
+  'live.watchers': '{count} watcher(s)',
+  'live.watch': 'Watch Game',
+  'live.view_spectator': 'Open Spectator Mode',
+  'live.mode_quick_play': 'Quick Play',
+  'live.mode_rated': 'Rated Quick Play',
+  'live.mode_casual': 'Casual Quick Play',
+  'live.moves': '{count} moves',
+  'live.updated': 'Last move {n}m ago',
+  'live.just_now': 'Updated just now',
+
+  // Feedback
+  'feedback.button': 'Feedback',
+  'feedback.title': 'Send Feedback',
+  'feedback.bug': 'Bug',
+  'feedback.feature': 'Feature',
+  'feedback.other': 'Other',
+  'feedback.placeholder_bug': 'What happened? What did you expect?',
+  'feedback.placeholder_feature': 'What feature would you like?',
+  'feedback.placeholder_other': 'Tell us anything...',
+  'feedback.thanks': 'Thank you!',
+  'feedback.thanks_desc': 'Your feedback helps improve ThaiChess Online.',
+  'feedback.error': 'Failed to send. Please try again.',
+
+  // Feedback Messages Page
+  'feedback_page.title': 'Feedback Messages',
+  'feedback_page.count': '{count} message(s)',
+  'feedback_page.filter_all': 'All',
+  'feedback_page.empty': 'No feedback yet',
+  'feedback_page.empty_desc': 'Feedback submitted by users will appear here.',
+  'feedback_page.detail_page': 'Page',
+  'feedback_page.detail_date': 'Date',
+  'feedback_page.loading': 'Loading moderation...',
+  'feedback_page.load_failed': 'Failed to load feedback.',
+  'feedback_page.moderate_failed': 'Failed to moderate feedback.',
+  'feedback_page.removed_by_admin': 'Removed by admin',
+  'feedback_page.delete': 'Delete',
+  'fair_play.report_action': 'Report Suspicious Play',
+  'fair_play.report_sent': 'Fair-play report sent for review.',
+  'fair_play.report_sent_short': 'Report Sent',
+  'fair_play.report_failed': 'Failed to submit fair-play report.',
+  'fair_play.admin_title': 'Fair Play Cases',
+  'fair_play.admin_desc': 'Review suspicious rated-play reports and restrict only after manual confirmation.',
+  'fair_play.admin_count': '{count} case(s)',
+  'fair_play.admin_empty': 'No fair-play cases match this filter.',
+  'fair_play.admin_load_failed': 'Failed to load fair-play cases.',
+  'fair_play.admin_action_failed': 'Failed to update fair-play case.',
+  'fair_play.admin_note_default': 'Reviewed by admin',
+  'fair_play.filter_all': 'All',
+  'fair_play.filter_open': 'Open',
+  'fair_play.filter_reviewed': 'Reviewed',
+  'fair_play.filter_restricted': 'Restricted',
+  'fair_play.filter_dismissed': 'Dismissed',
+  'fair_play.action_restrict': 'Restrict Rated',
+  'fair_play.action_clear': 'Clear Restriction',
+  'fair_play.action_dismiss': 'Dismiss',
+  'fair_play.event_count': '{count} signal(s)',
+  'fair_play.updated_at': 'Updated',
+  'fair_play.case_id': 'Case',
+  'fair_play.restricted_since': 'Restricted since',
+  'fair_play.status.open': 'Open',
+  'fair_play.status.reviewed': 'Reviewed',
+  'fair_play.status.restricted': 'Restricted',
+  'fair_play.status.dismissed': 'Dismissed',
+  'fair_play.event.analysis_blocked': 'Analysis blocked',
+  'fair_play.event.user_reported': 'User reported',
+  'fair_play.event.none': 'No event',
+
+  // Time ago
+  'time.just_now': 'just now',
+  'time.min_ago': '{n}m ago',
+  'time.hour_ago': '{n}h ago',
+  'time.day_ago': '{n}d ago',
+
+  // Analysis
+  'analysis.title': 'Quick Analysis',
+  'analysis.loading': 'Loading game...',
+  'analysis.parse_failed': 'Failed to parse game data',
+  'analysis.game_not_found': 'Game not found',
+  'analysis.no_moves': 'Game has no moves to analyze',
+  'analysis.failed': 'Analysis failed',
+  'analysis.analyzing': 'Reviewing game...',
+  'analysis.analyzing_note': 'Running Fairy-Stockfish review for this game. The report will appear here when the engine finishes.',
+  'analysis.progress': 'Move {current} of {total}',
+  'analysis.elapsed': 'Elapsed {seconds}s',
+  'analysis.slow_hint': 'Long reviews can take longer on the current server. The report will continue when the next position finishes.',
+  'analysis.engine_title': 'Review Source',
+  'analysis.engine_note': 'This review was completed with {engine}.',
+  'analysis.provisional_title': 'Provisional Review',
+  'analysis.provisional_note': 'This review used fallback analysis in some positions ({engine}), so the accuracy can still be overstated.',
+  'analysis.accuracy': 'Estimated Accuracy',
+  'analysis.eval_graph': 'Evaluation',
+  'analysis.show_best': 'Show suggested move',
+  'analysis.keyboard_hint': 'Use arrow keys to navigate moves',
+  'analysis.eval_before': 'Before',
+  'analysis.eval_after': 'After',
+  'analysis.best_was': 'Suggestion was',
+  'analysis.best': 'Best',
+  'analysis.excellent': 'Excellent',
+  'analysis.good': 'Good',
+  'analysis.inaccuracy': 'Inaccuracy',
+  'analysis.mistake': 'Mistake',
+  'analysis.blunder': 'Blunder',
+  'analysis.analyze': 'Analyze Game',
+  'analysis.view': 'View Analysis',
+  'analysis.next': 'Next',
+  'analysis.prev': 'Back',
+  'analysis.note': 'Engine review estimates move quality from changing winning chances, similar to Lichess-style accuracy.',
+  'analysis.move_accuracy': 'Move accuracy',
+  'analysis.desc_best': 'This matches the strongest engine continuation found in the review.',
+  'analysis.desc_good': 'A solid move that keeps most of your winning chances.',
+  'analysis.desc_inaccuracy': 'This move gives up some winning chances, but the position is still playable.',
+  'analysis.desc_mistake': 'This move drops a noticeable amount of winning chances. The engine found a stronger continuation.',
+  'analysis.desc_blunder': 'This move drops a large amount of winning chances. The engine found a much stronger continuation.',
+  'analysis.editor.error': 'Position analysis failed',
+  'analysis.editor.reset_board': 'Reset board',
+  'analysis.editor.clear_board': 'Clear board',
+  'analysis.editor.analyzing_position': 'Analyzing...',
+  'analysis.editor.analyze_position': 'Analyze position',
+  'analysis.editor.tools': 'Editor Tools',
+  'analysis.editor.move_pieces': 'Move pieces',
+  'analysis.editor.erase_square': 'Erase square',
+  'analysis.editor.position': 'Position',
+  'analysis.editor.copy_position': 'Copy position',
+  'analysis.editor.copy_link': 'Copy link',
+  'analysis.editor.engine': 'Engine',
+  'analysis.editor.eval': 'Eval',
+  'analysis.editor.best_move': 'Best move',
+  'analysis.editor.none': 'none',
+  'analysis.editor.source': 'Source',
+  'analysis.editor.depth': 'Depth',
+  'analysis.editor.pv': 'PV',
+  'analysis.editor.label': 'Editor',
+  'analysis.editor.turn_to_move': '{color} to move',
+
+  // Auth / Account
+  'auth.sign_in': 'Sign in',
+  'auth.sign_in_desc': 'Accounts are optional. Anonymous play still works.',
+  'auth.hero_title': 'Keep your rated games and profile in one place.',
+  'auth.hero_desc': 'Use a one-time code sent to your email. You can still play casually without an account, but signing in keeps your rating, username, and match history tied to you.',
+  'auth.benefit_optional': 'Anonymous play still works for quick casual games.',
+  'auth.benefit_rated': 'Rated play and account settings stay attached to your email.',
+  'auth.benefit_fast': 'No password to remember. Just email, code, and you are in.',
+  'auth.back_to_play': 'Back to play',
+  'auth.step_email': 'Email',
+  'auth.step_code': 'Code',
+  'auth.email': 'Email',
+  'auth.email_placeholder': 'you@example.com',
+  'auth.email_hint': 'We will send a one-time 6-digit code to this address.',
+  'auth.code': '6-digit code',
+  'auth.code_placeholder': '123456',
+  'auth.code_hint': 'The code usually arrives within a minute. Check spam if you do not see it.',
+  'auth.send_code': 'Send sign-in code',
+  'auth.sending_code': 'Sending code...',
+  'auth.code_sent_notice': 'Check your email for a 6-digit sign-in code.',
+  'auth.code_sent_to': 'Code sent to {email}',
+  'auth.verify_code': 'Verify code',
+  'auth.signing_in': 'Signing in...',
+  'auth.use_another_email': 'Use another email',
+  'auth.send_code_failed': 'Failed to send code.',
+  'auth.sign_in_failed': 'Failed to sign in.',
+  'account.title': 'Account',
+  'account.username': 'Username',
+  'account.username_placeholder': '3-20 letters, numbers, or underscores',
+  'account.rating': 'Rating',
+  'account.rated_games': 'Rated games',
+  'account.wins': 'Wins',
+  'account.losses': 'Losses',
+  'account.draws': 'Draws',
+  'account.saving': 'Saving...',
+  'account.save_profile': 'Save profile',
+  'account.profile_updated': 'Profile updated.',
+  'account.update_failed': 'Failed to update profile.',
+  'account.open_feedback': 'Open feedback moderation',
+  'account.open_fair_play': 'Open fair-play moderation',
+  'account.rated_restricted_title': 'Rated play restricted',
+  'account.rated_restricted_desc': 'This account can still sign in and play casual games, but rated matchmaking is disabled pending fair-play review.',
+  'account.sign_out': 'Sign out',
+  'account.puzzle_eyebrow': 'Puzzle progress',
+  'account.puzzle_title': 'Training overview',
+  'account.puzzle_desc': 'Your puzzle history now follows your account across devices.',
+  'account.puzzle_completed_label': 'Solved',
+  'account.puzzle_remaining_label': 'Remaining',
+  'account.puzzle_focus_label': 'Top theme',
+  'account.puzzle_focus_empty': 'No theme yet',
+  'account.puzzle_next_label': 'Continue with',
+  'account.puzzle_continue': 'Continue training',
+  'account.puzzle_all_done': 'You cleared the current set',
+  'account.puzzle_all_done_desc': 'Browse the full list again and replay your toughest lessons.',
+  'account.puzzle_last_played_label': 'Last played',
+  'account.puzzle_last_played_empty': 'Start a puzzle and it will show up here.',
+  'account.puzzle_last_played_meta': '{status} · {date}',
+  'account.puzzle_recent_label': 'Recent solves',
+  'account.puzzle_recent_empty': 'No solved puzzles yet.',
+  'account.puzzle_recent_meta': 'Solved {date}',
+  'account.puzzle_status_in_progress': 'In progress',
+  'account.puzzle_status_solved': 'Solved',
+  'account.puzzle_percent': '{percent}% of the current {total}-puzzle set complete',
+
+  // Leaderboard
+  'leaderboard.eyebrow': 'Rated Ladder',
+  'leaderboard.title': 'Leaderboard',
+  'leaderboard.desc': 'Top rated players from finished rated quick-play games.',
+  'leaderboard.count': '{count} ranked player(s)',
+  'leaderboard.empty': 'No rated players yet',
+  'leaderboard.empty_desc': 'Finish a rated quick-play game to appear here.',
+  'leaderboard.col_rank': 'Rank',
+  'leaderboard.col_player': 'Player',
+  'leaderboard.col_rating': 'Rating',
+  'leaderboard.col_games': 'Games',
+  'leaderboard.col_record': 'Record',
+  'leaderboard.col_winrate': 'Win Rate',
+  'leaderboard.you': 'You',
+  'leaderboard.view_recent': 'Recent Games',
+  'leaderboard.mobile_stats': '{games} games • {record} • {winrate}',
+  'leaderboard.load_failed': 'Failed to load leaderboard ({status})',
+
+  // Error UI
+  'error.board_display': 'Board display error',
+  'error.connection_title': 'Connection Error',
+  'error.connection_body': 'An error occurred while loading data',
+  'error.try_again': 'Try Again',
+  'error.something_wrong': 'Something went wrong',
+  'error.unexpected': 'The app encountered an unexpected error. This has been noted.',
+  'error.unknown': 'Unknown error',
+  'error.reload_page': 'Reload Page',
+  'error.report_bug': 'Report this bug on GitHub',
+
+  // Header
+  'header.admin': 'Admin',
+  'header.sign_in': 'Sign In',
+  'header.menu': 'Menu',
+  'header.close_menu': 'Close',
+  'header.switch_to_th': 'Switch to Thai',
+  'header.switch_to_en': 'Switch to English',
+
+  // Guide Page
+  'guide.cta_title': 'Ready to try a game?',
+  'guide.cta_desc': 'If you want a smooth first session, start with the bot or puzzles, then move into live games once the pieces feel natural.',
+  'guide.cta_next': 'Go to the next guide',
+
+  // Footer
+  'footer.links_label': 'Site footer links',
+  'footer.what_is_makruk': 'What Is Makruk?',
+  'footer.how_to_play_makruk': 'How to Play Makruk',
+
+  // Language
+  'lang.switch': 'TH',
+};
+
+const TH: Record<string, string> = {
+  // Common
+  'app.name': 'หมากรุก',
+  'app.tagline': 'หมากรุกไทยออนไลน์',
+  'nav.play': 'เล่น',
+  'nav.watch': 'ชมสด',
+  'nav.lessons': 'บทเรียน',
+  'nav.puzzles': 'ปริศนา',
+  'nav.games': 'เกม',
+  'nav.about': 'เกี่ยวกับ',
+  'common.back_home': 'กลับหน้าแรก',
+  'common.white': 'ขาว',
+  'common.black': 'ดำ',
+  'common.draw': 'เสมอ',
+  'common.you': 'คุณ',
+  'common.guest': 'ผู้เล่นทั่วไป',
+  'common.cancel': 'ยกเลิก',
+  'common.send': 'ส่ง',
+  'common.sending': 'กำลังส่ง...',
+  'common.close': 'ปิด',
+  'common.retry': 'ลองใหม่',
+  'common.new_game': 'เกมใหม่',
+  'common.loading': 'กำลังโหลด...',
+  'footer.tagline': 'ThaiChess — ฟรีและโอเพนซอร์ส',
+  'footer.inspired': 'ได้แรงบันดาลใจจาก',
+  'footer.community': 'ชุมชน',
+  'footer.github': 'GitHub',
+  'footer.star_github': 'ติดดาวบน GitHub',
+
+  // Appearance
+  'appearance.title': 'กระดานและตัวหมาก',
+  'appearance.board_and_pieces': 'กระดานและตัวหมาก',
+  'appearance.subtitle': 'เลือกธีมกระดานและธีมสีตัวหมากที่อ่านง่ายครั้งเดียว แล้วใช้ได้ทั่วทั้งเว็บ ทั้งออนไลน์ บอท เล่นในเครื่อง และปริศนา',
+  'appearance.open': 'กระดานและตัวหมาก',
+  'appearance.open_short': 'ธีม',
+  'appearance.boards_tab': 'กระดาน',
+  'appearance.colors_tab': 'สีตัวหมาก',
+  'appearance.live_preview': 'พรีวิวสด',
+  'appearance.preview_title': 'ตัวอย่าง',
+  'appearance.preview_subtitle': 'เลื่อนเมาส์เพื่อดูตัวอย่าง คลิกเพื่อใช้งานทันที ระบบจะบันทึกไว้ในอุปกรณ์นี้อัตโนมัติ',
+  'appearance.preview_board': 'ธีมกระดาน',
+  'appearance.preview_piece_theme': 'ธีมสีตัวหมาก',
+  'appearance.core_shape_label': 'รูปทรงหลัก',
+  'appearance.core_shape_desc': 'เป็นรูปทรงเดียวที่รองรับในการเล่นจริง เพื่อให้มองออกเร็วและสม่ำเสมอทุกเกม',
+  'appearance.saved_note': 'บันทึกอัตโนมัติในอุปกรณ์นี้',
+  'appearance.category_classic': 'คลาสสิก',
+  'appearance.category_soft': 'นุ่มนวล',
+  'appearance.category_dark': 'โหมดกลางคืน',
+  'appearance.category_elegant': 'เรียบหรู',
+  'appearance.readability_checked': 'ผ่านการเช็กความชัด',
+  'appearance.grid_contrast': 'คอนทราสต์เส้นตาราง',
+  'appearance.piece_contrast': 'คอนทราสต์ตัวหมากต่ำสุด',
+  'appearance.single_surface_label': 'กระดานสีเดียวแบบหมากรุกไทย',
+  'appearance.single_surface_desc': 'ธีมจะเปลี่ยนเฉพาะโทนกระดาน ผิวสัมผัส และเส้นตารางเท่านั้น',
+  'appearance.comparison_title': 'แบบเดิมเทียบกับหมากรุกไทย',
+  'appearance.comparison_desc': 'ธีมจะไม่สลับสีช่องอีกต่อไป แต่จะเปลี่ยนเฉพาะวัสดุและเส้นตารางของกระดาน',
+  'appearance.comparison_before': 'ก่อนปรับ',
+  'appearance.comparison_after': 'หมากรุกไทยตอนนี้',
+
+  // Home Page
+  'home.hero_title': 'เล่นหมากรุกได้ทันที',
+  'home.hero_desc': 'ไม่ต้องสมัคร เริ่มเกมหมากรุกไทยได้ในไม่กี่วินาที เล่นกับเพื่อน หรือฝึกกับบอท',
+  'home.quick_play': 'เล่นเลย',
+  'home.quick_play_desc': 'จับคู่ทันทีสำหรับเกมสบาย ๆ หรือเกมจัดอันดับ',
+  'home.find_opponent': 'หาคู่แข่ง',
+  'home.no_signup': 'ไม่ต้องสมัคร',
+  'home.free_to_play': 'เล่นฟรี',
+  'home.games_played': 'เล่นแล้ว {count} เกม',
+  'home.play_friend': 'เล่นกับเพื่อน',
+  'home.play_friend_desc': 'แชร์ลิงก์',
+  'home.play_bot': 'เล่นกับบอท',
+  'home.play_bot_desc': '10 บุคลิกบอทธีมเฉพาะ',
+  'home.play_bot_long_desc': 'ท้าดวลกับคู่แข่งหมากรุกไทยที่มีเอกลักษณ์ ชื่อ เรตติ้ง และสไตล์การเล่นต่างกันชัดเจน',
+  'home.puzzles': 'ปริศนา',
+  'home.puzzles_desc': 'ฝึกกลยุทธ์',
+  'home.puzzles_long_desc': 'ฝึกแทคติกด้วยปริศนาที่คัดมาจากเกมจริง',
+  'home.lessons': 'บทเรียน',
+  'home.lessons_desc': 'คอร์สหมากรุกไทยแบบเป็นลำดับ',
+  'home.create_private': 'สร้างเกมส่วนตัว',
+  'home.private_desc': 'เลือกเวลา เลือกสี แล้วแชร์ลิงก์เกมให้เพื่อน',
+  'home.time_control': 'เวลาเล่น',
+  'home.choose_color': 'เลือกสี',
+  'home.color_random': 'สุ่ม',
+  'home.color_white': 'ขาว',
+  'home.color_black': 'ดำ',
+  'home.play_with_friend': 'เล่นกับเพื่อน',
+  'home.creating': 'กำลังสร้าง...',
+  'home.or': 'หรือ',
+  'home.play_local': 'เล่นในเครื่อง (จอเดียวกัน)',
+  'home.play_local_desc': 'สลับกันเดินบนกระดานเดียวในหน้าจอเดียว',
+  'home.watch_live': 'ชมเกมสด',
+  'home.watch_live_desc': 'ดูเกมหมากรุกไทยสาธารณะที่กำลังเล่น และเข้าโหมดผู้ชมได้ทันที',
+  'home.live_now_title': 'กำลังเล่นตอนนี้',
+  'home.live_now_desc': 'เกมสาธารณะที่กำลังเล่นจะอัปเดตแบบเรียลไทม์ เปิดกระดานไหนก็ได้ในโหมดผู้ชมแบบอ่านอย่างเดียว',
+  'home.view_all_live': 'ดูเกมสดทั้งหมด',
+  'home.no_live_games': 'ตอนนี้ยังไม่มีเกมสด',
+  'home.no_live_games_desc': 'กลับมาดูอีกครั้งเมื่อมีเกมสาธารณะกำลังเล่น',
+  'home.join_prompt': 'มีรหัสเกม?',
+  'home.join_link': 'เข้าร่วมเกม',
+  'home.join_title': 'เข้าร่วมเกม',
+  'home.join_desc': 'เปิดเกมที่เพื่อนแชร์มาได้ทันทีด้วยรหัสเกม',
+  'home.join_placeholder': 'ใส่รหัสเกม...',
+  'home.join': 'เข้าร่วม',
+  'home.training_start': 'เริ่มชุดปริศนา',
+  'home.training_continue': 'ฝึกต่อ',
+  'home.training_progress': 'สำเร็จแล้ว {done}/{total} ปริศนา',
+  'home.training_focus': 'ธีมที่เด่นสุดตอนนี้: {theme}',
+  'home.training_resume': 'เล่นล่าสุด: {title}',
+  'home.training_recent': 'แก้ล่าสุด: {title}',
+  'home.streak_start': 'เริ่มสตรีค',
+  'home.streak_continue': 'ลุยต่อ',
+  'home.streak_title': 'Puzzle Streak',
+  'home.streak_progress': 'แก้บทเรียนแล้ว {done}/{total} ข้อ',
+  'home.streak_focus': 'ธีมบทเรียนที่เด่นตอนนี้: {theme}',
+  'home.streak_resume': 'บทเรียนล่าสุดที่เล่น: {title}',
+  'home.streak_recent': 'บทเรียนล่าสุดที่แก้ได้: {title}',
+
+  // Time presets
+  'time.bullet': 'บุลเลต',
+  'time.blitz': 'บลิตซ์',
+  'time.rapid': 'ราปิด',
+  'time.classical': 'คลาสสิก',
+
+  // Bot Game
+  'bot.title': 'เล่นกับบอท',
+  'bot.setup_title': 'เล่นกับคอมพิวเตอร์',
+  'bot.setup_desc': 'เลือกบุคลิกบอท ดูสไตล์ ระดับ และพลังฝีมือโดยประมาณ เลือกฝั่ง แล้วเริ่มเกมซ้อมได้เลย หากเลือกสุ่ม ระบบจะตัดสินฝั่งให้ตอนเริ่มเกมเท่านั้น',
+  'bot.difficulty': 'ระดับความยาก',
+  'bot.level_label': 'เลเวล',
+  'bot.level_short': 'เลเวล {level}',
+  'bot.estimated_elo_label': 'ELO โดยประมาณ',
+  'bot.estimated_elo_range': 'ประมาณ {range} ELO',
+  'bot.estimated_elo_note': 'ค่าพลังฝีมือโดยประมาณนี้อ้างอิงจากพฤติกรรมการเล่น ไม่ใช่เรตทางการ',
+  'bot.easy': 'ง่าย',
+  'bot.easy_desc': 'เดินสุ่มกินตัวพื้นฐาน',
+  'bot.medium': 'ปานกลาง',
+  'bot.medium_desc': 'คิดล่วงหน้า 2 ตา',
+  'bot.hard': 'ยาก',
+  'bot.hard_desc': 'คิดล่วงหน้า 4 ตา',
+  'bot.play_as': 'เลือกฝั่ง',
+  'bot.random': 'สุ่ม',
+  'bot.random_assigned': 'กำหนดตอนเริ่มเกม',
+  'bot.side_locked': 'ล็อกฝั่งก่อนเริ่มเกม',
+  'bot.start': 'เริ่มเกม',
+  'bot.thinking': 'กำลังคิด...',
+  'bot.your_turn': 'ตาของคุณ',
+  'bot.bot_thinking': 'บอทกำลังคิด...',
+  'bot.you_won': 'คุณชนะ!',
+  'bot.bot_wins': 'บอทชนะ',
+  'bot.resign': 'ยอมแพ้',
+  'bot.resign_confirm': 'คุณแน่ใจหรือไม่ว่าต้องการยอมแพ้?',
+  'bot.vs_bot': 'กับบอท',
+
+  // Puzzles
+  'puzzle.title': 'Puzzle Streak',
+  'puzzle.desc': 'เล่นปริศนาต่อเนื่อง เก็บคะแนน และให้ระบบปรับความยากแบบซ่อนอยู่ตามผลงานของคุณ',
+  'puzzle.completed': '{done}/{total} สำเร็จ',
+  'puzzle.completed_summary': 'ความคืบหน้าทั้งหมด',
+  'puzzle.progress_hint': 'ปริศนาที่แก้แล้วจะถูกจำไว้ในอุปกรณ์นี้',
+  'puzzle.all': 'ทั้งหมด',
+  'puzzle.all_lessons': 'บทเรียนทั้งหมด',
+  'puzzle.beginner': 'เริ่มต้น',
+  'puzzle.intermediate': 'ปานกลาง',
+  'puzzle.advanced': 'ขั้นสูง',
+  'puzzle.filter_label': 'ชุดฝึกตอนนี้',
+  'puzzle.filter_all_desc': 'รวมทุกบทเรียนที่เปิดใช้ไว้ในชุดเดียว',
+  'puzzle.filter_beginner_desc': 'แท็กติกสั้นและรูปจนพื้นฐาน',
+  'puzzle.filter_intermediate_desc': 'แนวคิดสองจังหวะที่ลงโทษเกมรับหลวม',
+  'puzzle.filter_advanced_desc': 'โจทย์คำนวณหนักขึ้นเมื่อมีชุดเพิ่ม',
+  'puzzle.theme_drill_title': 'ฝึกตามธีม',
+  'puzzle.theme_drill_desc': 'เลือกฝึกแนวคิดเดียวซ้ำ ๆ เมื่ออยากซ้อมเฉพาะทาง',
+  'puzzle.theme_drill_count': '{count} ธีมในชุดนี้',
+  'puzzle.theme_all': 'ทุกแนวคิด',
+  'puzzle.track_progress': 'ความคืบหน้าชุด {track}',
+  'puzzle.track_completed': '{done}/{total} ในชุดนี้',
+  'puzzle.remaining_label': 'ยังไม่เคยแก้',
+  'puzzle.remaining_desc': 'โจทย์ใหม่จะถูกจัดไว้ก่อนเสมอ',
+  'puzzle.focus_label': 'ธีมเด่น',
+  'puzzle.focus_desc': 'ช่วยเลือกว่าจะฝึกแนวคิดไหนต่อ',
+  'puzzle.focus_empty': 'ยังไม่มีธีม',
+  'puzzle.next_up': 'บทเรียนถัดไป',
+  'puzzle.next_up_fresh': 'เริ่มจากโจทย์ที่ยังไม่เคยแก้ในชุดนี้ก่อน',
+  'puzzle.next_up_review': 'คุณเคลียร์ชุดนี้แล้ว ลองทบทวนโจทย์เดิมเพื่อให้แม่นขึ้น',
+  'puzzle.start_here': 'เริ่มตรงนี้',
+  'puzzle.practice_title': 'คิวฝึก',
+  'puzzle.practice_desc': 'โจทย์ใหม่จะมาก่อน ส่วนโจทย์ที่แก้แล้วเก็บไว้ทบทวน',
+  'puzzle.new_badge': 'ใหม่',
+  'puzzle.solved_badge': 'แก้แล้ว',
+  'puzzle.empty_title': 'ยังไม่มีปริศนาในชุดนี้',
+  'puzzle.empty_desc': 'ลองเปลี่ยนระดับความยาก หรือกลับมาอีกครั้งเมื่อมีบทเรียนเพิ่ม',
+  'puzzle.empty_theme_title': 'ยังไม่มีโจทย์ธีม {theme} ในชุดนี้',
+  'puzzle.empty_theme_desc': 'ลองเปลี่ยนแนวคิด หรือออกจากชุด {track} เพื่อหาโจทย์ {theme} เพิ่ม',
+  'puzzle.to_move': '{color} เดิน',
+  'puzzle.find_best': 'หาตาเดินที่ดีที่สุดสำหรับฝั่ง{color}!',
+  'puzzle.step': 'ขั้นที่ {current} จาก {total}',
+  'puzzle.correct': 'ถูกต้อง!',
+  'puzzle.solved_hint': 'แก้ได้โดยใช้คำใบ้',
+  'puzzle.solved_clean': 'แก้ได้โดยไม่ใช้คำใบ้!',
+  'puzzle.wrong': 'ไม่ถูกต้อง!',
+  'puzzle.wrong_desc': 'นั่นไม่ใช่ตาเดินที่ดีที่สุด ลองอีกครั้ง!',
+  'puzzle.hint': 'คำใบ้',
+  'puzzle.next': 'ปริศนาถัดไป',
+  'puzzle.previous': 'ก่อนหน้า',
+  'puzzle.all_puzzles': 'ปริศนาทั้งหมด',
+  'puzzle.back_to_lessons': 'กลับไปบทเรียน',
+  'puzzle.not_found': 'ไม่พบปริศนา',
+  'puzzle.back': 'กลับไปปริศนา',
+  'puzzle.lesson': 'บทเรียน',
+  'puzzle.lessons_nav': 'บทเรียน',
+  'puzzle.lessons_eyebrow': 'โหมดบทเรียน',
+  'puzzle.lessons_title': 'บทเรียน',
+  'puzzle.lessons_desc': 'เรียนหมากรุกไทยเป็นคอร์สตามลำดับ แล้วค่อยต่อยอดด้วยปริศนาที่ตรงกับแนวคิดของบทเรียน',
+  'puzzle.lessons_tracks_title': 'เส้นทางการเรียน',
+  'puzzle.lessons_tracks_desc': 'Beginner, Intermediate และ Advanced ยังคงอยู่ที่นี่ในฐานะเส้นทางฝึกแบบตั้งใจ',
+  'puzzle.play_streak': 'เล่นสตรีค',
+  'puzzle.streak_nav': 'Puzzle Streak',
+  'puzzle.streak_eyebrow': 'โหมดหลัก',
+  'puzzle.streak_title': 'Puzzle Streak',
+  'puzzle.streak_desc': 'แก้ปริศนาต่อเนื่อง เก็บสตรีค และให้ระบบค่อย ๆ ปรับความท้าทายขึ้นหรือลงจากการเล่นของคุณ',
+  'puzzle.streak_prompt_title': 'รักษาสตรีคไว้',
+  'puzzle.streak_prompt_desc': 'ตอบถูกแล้วได้คะแนนทันที และโจทย์ถัดไปจะขึ้นต่อแบบลื่นไหล',
+  'puzzle.streak_score_label': 'คะแนน',
+  'puzzle.streak_label': 'สตรีค',
+  'puzzle.streak_session_label': 'แก้ได้',
+  'puzzle.streak_checkpoint_label': 'เช็กพอยต์',
+  'puzzle.streak_checkpoint_progress': '{current}/{total} ถึงสัญญาณถัดไป',
+  'puzzle.streak_flow_label': 'จังหวะ',
+  'puzzle.streak_flow_desc': 'ปรับแบบอัตโนมัติ',
+  'puzzle.streak_puzzle_label': 'โจทย์ {number}',
+  'puzzle.streak_points': '+{points} คะแนน',
+  'puzzle.streak_focus_title': 'โจทย์ตอนนี้',
+  'puzzle.streak_focus_keep': 'ตั้งหลักใหม่ แล้วเริ่มอีกครั้ง',
+  'puzzle.streak_focus_desc': 'โหมดนี้ซ่อนความยากไว้ เพื่อให้เล่นต่อได้เร็วและลื่นที่สุด',
+  'puzzle.streak_session_title': 'เซสชัน',
+  'puzzle.streak_best_label': 'สตรีคสูงสุด',
+  'puzzle.streak_lessons_progress_label': 'บทเรียน',
+  'puzzle.visit_lessons': 'เปิดบทเรียน',
+  'puzzle.streak_end_eyebrow': 'จบเซสชัน',
+  'puzzle.streak_end_title': 'สตรีคจบแล้ว',
+  'puzzle.streak_end_desc': 'พลาดหนึ่งครั้งก็จบสตรีค แต่ความคืบหน้าของบทเรียนยังถูกบันทึกไว้',
+  'puzzle.streak_ended_at': 'สตรีคจบที่ {streak}',
+  'puzzle.streak_end_summary': 'สตรีคสุดท้าย {streak} · คะแนนเซสชัน {score}',
+  'puzzle.retry_puzzle': 'ลองโจทย์เดิมอีกครั้ง',
+  'puzzle.streak_try_again': 'เริ่มสตรีคใหม่',
+  'puzzle.streak_milestone_improving': 'คุณกำลังดีขึ้น เล่นต่อให้ลื่นเลย',
+  'puzzle.streak_milestone_harder': 'ตอนนี้กำลังเจอโจทย์ที่ยากขึ้นแล้ว ระวังให้ดี',
+  'puzzle.source': 'ที่มา',
+  'puzzle.source_real_game_ply': 'เกมจริง · ply {ply}',
+  'puzzle.source_seed_game_ply': 'เกมตัวอย่าง · ply {ply}',
+  'puzzle.source_starter_pack': 'ชุดเริ่มต้น',
+  'puzzle.source_curated_tactic': 'แทคติกคัดสรร',
+  'puzzle.source_tactical_motif': 'มอติฟแทคติก',
+  'puzzle.source_review_batch': 'ชุดรอตรวจ',
+  'puzzle.rating_label': 'ระดับที่แนะนำ',
+  'puzzle.rating_short': 'เรต {score}',
+  'puzzle.success_rate': 'อัตราสำเร็จ',
+  'puzzle.attempts_label': 'จำนวนครั้ง',
+  'puzzle.activity_title': 'ความคืบหน้ารอบนี้',
+  'puzzle.activity_status_label': 'สถานะ',
+  'puzzle.activity_status_new': 'ใหม่',
+  'puzzle.activity_status_in_progress': 'กำลังฝึก',
+  'puzzle.activity_status_solved': 'แก้แล้ว',
+  'puzzle.activity_last_played': 'เล่นล่าสุด {date}',
+  'puzzle.activity_completed_on': 'แก้สำเร็จ {date}',
+  'puzzle.related_theme_title': 'ธีมเดียวกันต่อ',
+  'puzzle.related_theme_desc': 'ต่อยอดธีม {theme} ด้วยโจทย์ถัดไปเหล่านี้',
+  'puzzle.related_theme_empty': 'ยังไม่มีโจทย์ {theme} เพิ่ม',
+
+  // Puzzle themes
+  'theme.BasicCheckmate': 'รูปจนพื้นฐาน',
+  'theme.BackRank': 'รุกจนแถวหลัง',
+  'theme.Checkmate': 'รุกจน',
+  'theme.Clearance': 'เปิดทาง',
+  'theme.Decoy': 'ล่อ',
+  'theme.Defense': 'ตั้งรับ',
+  'theme.BestDefense': 'ตั้งรับที่ดีที่สุด',
+  'theme.Deflection': 'เบี่ยงตัวป้องกัน',
+  'theme.CountDefense': 'กันเสมอด้วยการนับ',
+  'theme.CountingDraw': 'เสมอเพราะการนับ',
+  'theme.DoubleAttack': 'โจมตีสองเป้า',
+  'theme.DoubleCheck': 'รุกสองทาง',
+  'theme.ExchangeSacrifice': 'สังเวยเรือแลกตัว',
+  'theme.HangingPiece': 'ตัวลอย',
+  'theme.Interference': 'ขวางแนว',
+  'theme.MateIn1': 'จนใน 1',
+  'theme.MateIn2': 'จนใน 2',
+  'theme.MateIn3': 'จนใน 3+',
+  'theme.MatingNet': 'ข่ายจน',
+  'theme.Overloading': 'รับหลายหน้าที่',
+  'theme.PerpetualCheck': 'รุกไล่ไม่จบ',
+  'theme.Fork': 'โจมตีซ้อน',
+  'theme.Pin': 'ตรึง',
+  'theme.Promotion': 'หงายเบี้ย',
+  'theme.Discovery': 'เปิดแนว',
+  'theme.RemovalOfDefender': 'กำจัดตัวป้องกัน',
+  'theme.OnlyMove': 'ตาเดียวเท่านั้น',
+  'theme.Sacrifice': 'สังเวย',
+  'theme.SaveTheDraw': 'หาทางเสมอ',
+  'theme.Simplification': 'ผ่อนเข้าสู่ชนะง่าย',
+  'theme.SmotheredMate': 'ม้าจนปิดล้อม',
+  'theme.Skewer': 'เสียบ',
+  'theme.Stalemate': 'อับจนเสมอ',
+  'theme.SupportMate': 'จนมีตัวค้ำ',
+  'theme.Tactic': 'กลยุทธ์',
+  'theme.TacticalWin': 'ชนะเชิงกลยุทธ์',
+  'theme.TrappedPiece': 'ตัวติดกับ',
+  'theme.VulnerableKing': 'กษัตริย์เปราะบาง',
+  'theme.WinBeforeCountExpires': 'ชนะก่อนหมดนับ',
+  'theme.Windmill': 'กังหันกิน',
+  'theme.XRay': 'โจมตีทะลุแนว',
+  'theme.Zwischenzug': 'แทรกหมาก',
+  'theme.Endgame': 'เกมจบ',
+  'theme.Desperado': 'ตายก่อนขอกิน',
+
+  // Quick Play / Matchmaking
+  'quick.title': 'เล่นด่วน',
+  'quick.desc': 'หาคู่แข่งทันที ไม่ต้องแชร์ลิงก์!',
+  'quick.searching': 'กำลังหาคู่แข่ง...',
+  'quick.search_time': 'กำลังค้นหามา {time}',
+  'quick.queue': '{count} ผู้เล่นในคิว',
+  'quick.find': 'หาคู่แข่ง',
+  'quick.rated_available': 'จัดอันดับได้',
+  'quick.rated_unavailable': 'ปิดเกมจัดอันดับ',
+  'quick.casual_only': 'เกมทั่วไป',
+  'quick.rated_signed_in': 'จะเป็นเกมจัดอันดับเมื่อเจอคู่แข่งที่เข้าสู่ระบบเช่นกัน',
+  'quick.rated_restricted': 'บัญชีนี้ยังเล่นควิกเพลย์แบบทั่วไปได้ แต่ปิดการจับคู่แบบจัดอันดับไว้',
+  'quick.rated_sign_in': 'เข้าสู่ระบบเพื่อปลดล็อกเกมจัดอันดับ',
+
+  // Game Page
+  'game.waiting_title': 'รอคู่แข่ง',
+  'game.waiting_desc': 'แชร์ลิงก์นี้ให้เพื่อนเพื่อเริ่มเล่น',
+  'game.copy': 'คัดลอก',
+  'game.copied': 'คัดลอกแล้ว!',
+  'game.playing_as': 'คุณเล่นฝั่ง{color}',
+  'game.rated': 'จัดอันดับ',
+  'game.casual': 'ทั่วไป',
+  'game.rating_change': 'เรตติ้ง',
+  'game.connecting': 'กำลังเชื่อมต่อ...',
+  'game.error': 'ข้อผิดพลาด',
+  'game.opponent': 'คู่แข่ง',
+  'game.online': 'ออนไลน์',
+  'game.offline': 'ออฟไลน์',
+  'game.active_now': 'กำลังเล่น',
+  'game.idle': 'ไม่ได้ใช้งาน',
+  'game.away': 'อยู่นอกหน้าเกม',
+  'game.ping_short': 'ปิง',
+  'game.ping_unknown': 'ไม่มีข้อมูลปิง',
+  'game.ping_value': 'ค่าหน่วง {ms} ms',
+  'game.to_move': 'ถึงตา',
+  'game.your_turn': 'ตาของคุณ',
+  'game.opponent_turn': 'ตาฝ่ายตรงข้าม',
+  'game.you_won': 'คุณชนะ!',
+  'game.you_lost': 'คุณแพ้',
+  'game.draw_offer_received': 'คู่แข่งเสนอเสมอ',
+  'game.accept': 'ยอมรับ',
+  'game.decline': 'ปฏิเสธ',
+  'game.offer_draw': '½ เสมอ',
+  'game.resign': '⚐ ยอมแพ้',
+  'game.resign_confirm': 'คุณแน่ใจหรือไม่ว่าต้องการยอมแพ้?',
+  'game.share': 'แชร์',
+  'game.open_spectator': 'เปิดมุมมองผู้ชม',
+  'game.piece_guide': '📖 คู่มือตัวหมาก',
+  'game.opponent_dc': 'คู่แข่งหลุดออก กำลังรอเชื่อมต่อใหม่...',
+  'game.game_label': 'เกม:',
+  'game.counting_title': 'กฎการนับ',
+  'game.counting_board_honor': 'การนับแบบหมดเบี้ย',
+  'game.counting_pieces_honor': 'การนับแบบขุนเปล่า',
+  'game.counting_status': '{type}: ฝั่ง{color} นับอยู่ที่ {current}/{limit}',
+  'game.counting_final': '{type}: ตาโจมตีสุดท้ายก่อนเสมอ',
+  'game.counting_available': '{type}: ฝั่ง{color} สามารถเริ่มนับได้',
+  'game.counting_start': 'เริ่มนับ',
+  'game.counting_stop': 'หยุดนับ',
+  'game.premove_set': 'ตั้งตาเดินล่วงหน้าแล้ว',
+  'game.check_status': 'กำลังถูกรุก',
+  'game.nav_hint': 'ใช้ปุ่มลูกศรเพื่อดูตาเดิน',
+  'game.select_piece_style': 'เลือกรูปแบบตัวหมาก',
+  'game.piece_style': 'ตัวหมาก',
+  'game.piece_style_current': 'ปัจจุบัน',
+  'game.piece_style_western': 'สากล',
+  'game.piece_style_makruk': 'หมากรุก',
+  'game.match_status': 'สถานะเกม',
+  'game.playing_as_label': 'คุณเล่นฝั่ง',
+  'game.review_mode': 'มุมมองกระดาน',
+  'game.reviewing_history': 'กำลังดูประวัติเดิน',
+  'game.live_position': 'ตำแหน่งปัจจุบัน',
+  'game.return_to_live': 'กลับไปตาเดินปัจจุบัน',
+  'game.live_match_title': 'เกมหมากรุกสด',
+  'game.live_match_desc': 'ให้กระดานเป็นจุดสำคัญที่สุด ส่วนเวลา การทบทวน และคำสั่งควรอยู่ใกล้มือโดยไม่แย่งความสนใจ',
+  'game.spectator_mode': 'โหมดผู้ชม',
+  'game.watching_live_game': 'กำลังชมเกมสด',
+  'game.spectator_desc': 'มุมมองแบบอ่านอย่างเดียว พร้อมกระดานสด นาฬิกา และรายการตาเดิน',
+  'game.live_now': 'สด',
+  'game.final_position': 'ตำแหน่งสุดท้าย',
+  'game.read_only': 'อ่านอย่างเดียว',
+  'game.flip_board': 'สลับมุมกระดาน',
+  'game.white_perspective': 'มุมมองฝ่ายขาว',
+  'game.black_perspective': 'มุมมองฝ่ายดำ',
+  'game.live_position_locked': 'ระหว่างเกม กระดานจะแสดงตำแหน่งสดล่าสุดเสมอ',
+  'game.spectator_turn': 'ถึงตาฝั่ง{color}',
+  'game.spectator_controls_hidden': 'ซ่อนคำสั่งสำหรับผู้เล่นในโหมดผู้ชม',
+  'game.post_game_title': 'ทบทวนเกมที่จบแล้ว',
+  'game.post_game_desc': 'ไล่ดูตาเดิน ศึกษาจุดจบของเกม หรือไปวิเคราะห์ต่อได้ทันที',
+  'game.side_title': 'แผงข้อมูลเกม',
+  'game.you_label': 'คุณ',
+  'game.reviewing_position': 'เกมจบแล้ว',
+
+  // Game Over Modal
+  'gameover.you_win': 'คุณชนะ!',
+  'gameover.you_lost': 'คุณแพ้',
+  'gameover.draw': 'เสมอ',
+  'gameover.by_checkmate': 'โดยรุกจน',
+  'gameover.by_resign': 'โดยยอมแพ้',
+  'gameover.by_timeout': 'หมดเวลา',
+  'gameover.by_stalemate': 'โดยอับ',
+  'gameover.by_agreement': 'ตกลงเสมอ',
+  'gameover.by_material': 'ตัวหมากไม่เพียงพอ',
+  'gameover.by_counting': 'โดยกฎการนับ',
+  'gameover.rematch': 'แข่งอีกครั้ง',
+  'gameover.rematch_accept': 'ยอมรับแข่งอีกครั้ง',
+  'gameover.rematch_sent': 'ส่งคำขอแล้ว',
+  'gameover.rematch_waiting': 'ส่งคำขอแข่งอีกครั้งแล้ว กำลังรอคู่แข่งตอบรับ',
+  'gameover.rematch_incoming': 'คู่แข่งต้องการแข่งอีกครั้ง',
+  'gameover.is_victorious': 'ชนะ',
+
+  // Local Game
+  'local.title': 'เล่นในเครื่อง',
+  'local.view_as': 'มุมมอง:',
+  'local.turn': 'ตาฝั่ง{color}',
+  'local.wins': 'ฝั่ง{color}ชนะ!',
+  'local.play_online': 'เล่นออนไลน์',
+
+  // Move History
+  'moves.title': 'ตาเดิน',
+  'moves.empty': 'ยังไม่มีตาเดิน',
+  'moves.first_position': 'ตำแหน่งแรก',
+  'moves.previous_move': 'ตาก่อนหน้า',
+  'moves.next_move': 'ตาถัดไป',
+  'moves.last_move': 'ตำแหน่งล่าสุด',
+
+  // Piece Guide
+  'guide.title': 'คู่มือตัวหมาก',
+  'guide.king': 'ขุน (King)',
+  'guide.queen': 'เม็ด (Queen)',
+  'guide.bishop': 'โคน (Bishop)',
+  'guide.rook': 'เรือ (Rook)',
+  'guide.knight': 'ม้า (Knight)',
+  'guide.pawn': 'เบี้ย (Pawn)',
+  'guide.promoted': 'เบี้ยหงาย (Promoted)',
+  'guide.king_move': 'เดินได้ 1 ช่องทุกทิศทาง',
+  'guide.queen_move': 'เดินได้ 1 ช่องแนวทแยง',
+  'guide.bishop_move': 'เดินได้ 1 ช่องแนวทแยงหรือ 1 ช่องหน้า',
+  'guide.rook_move': 'เดินได้ไม่จำกัดในแนวตั้งและแนวนอน',
+  'guide.knight_move': 'รูปตัว L: 2+1 ช่อง',
+  'guide.pawn_move': 'เดินหน้า 1 ช่อง; กินแนวทแยง',
+  'guide.promoted_move': 'เหมือนเม็ด (1 ช่องทแยง)',
+
+  // Connection
+  'conn.connected': 'เชื่อมต่อแล้ว',
+  'conn.reconnecting': 'กำลังเชื่อมต่อใหม่...',
+  'conn.disconnected': 'หลุดการเชื่อมต่อ — กำลังเชื่อมต่อใหม่...',
+
+  // About Page
+  'about.eyebrow': 'เรียนรู้หมากรุก',
+  'about.title': 'เกี่ยวกับหมากรุกไทย',
+  'about.hero_desc': 'Makruk คือหมากรุกดั้งเดิมของไทย จังหวะเกมค่อยเป็นค่อยไปกว่าเดิม ลดการแลกง่าย ๆ และเต็มไปด้วยรูปแบบตัวหมากที่ไม่เหมือนหมากรุกสากล ThaiChess ถูกสร้างขึ้นเพื่อให้คนเรียนรู้และเล่นออนไลน์ได้ง่ายขึ้น',
+  'about.cta_play': 'เริ่มเล่น',
+  'about.cta_puzzles': 'ลองปริศนา',
+  'about.quick_facts': 'ข้อมูลเร็ว',
+  'about.fact_heritage_label': 'ต้นกำเนิด',
+  'about.fact_heritage_value': 'หมากรุกไทยดั้งเดิม',
+  'about.fact_style_label': 'ลักษณะเกม',
+  'about.fact_style_value': 'รอบคอบ วางแผน ละเอียด',
+  'about.intro_title': 'ทำไมเราสร้างเว็บนี้',
+  'about.intro1': 'หมากรุกไทยเป็นแพลตฟอร์ม<strong>ฟรีและโอเพนซอร์ส</strong>สำหรับเล่นหมากรุกไทยกับทุกคนในโลก ไม่ต้องสมัคร ไม่มีโฆษณา ไม่มีค่าใช้จ่าย — แค่หมากรุกไทยบริสุทธิ์',
+  'about.intro2': 'พันธกิจของเราคือการทำให้<strong>หมากรุกไทยเป็นที่รู้จักทั่วโลก</strong> หมากรุกไทยเป็นหนึ่งในเกมกระดานที่เก่าแก่ที่สุดในโลก ใกล้เคียงกับจตุรงค์ของอินเดียมากกว่าหมากรุกสากล สมควรมีแพลตฟอร์มออนไลน์ระดับโลก',
+  'about.intro3': 'ได้แรงบันดาลใจจาก <a href="https://lichess.org" target="_blank" rel="noopener" class="text-primary hover:text-primary-light underline">Lichess</a> และ <a href="https://chess.com" target="_blank" rel="noopener" class="text-primary hover:text-primary-light underline">Chess.com</a> เราเชื่อว่าแพลตฟอร์มเกมดีๆ ควรเข้าถึงได้ทุกคน',
+  'about.stats_title': 'สถิติแพลตฟอร์ม',
+  'about.games_played': 'เกมที่เล่น',
+  'about.total_moves': 'ตาเดินทั้งหมด',
+  'about.white_wins': 'ฝั่งขาวชนะ',
+  'about.black_wins': 'ฝั่งดำชนะ',
+  'about.what_title': 'หมากรุกไทยคืออะไร?',
+  'about.what1': '<strong>หมากรุกไทย (Makruk)</strong> เป็นหมากรุกดั้งเดิมของไทย พัฒนามาจากจตุรงค์ของอินเดียโบราณ และเล่นกันในไทยมาหลายศตวรรษ',
+  'about.what2': 'ต่างจากหมากรุกสากล หมากรุกไทยยังคงรักษาลักษณะดั้งเดิมไว้ — เม็ด (ควีน) อ่อนแอ โคน (บิชอป) เดินต่างออกไป และเบี้ยเริ่มที่แถวที่สาม ทำให้เกมมีกลยุทธ์ลึกซึ้งและสวยงามเป็นเอกลักษณ์',
+  'about.what3': 'หมากรุกไทยเป็นที่นิยมในไทยและกัมพูชา (เรียกว่า อุกจตุรงค์) ได้รับการยอมรับจากสหพันธ์หมากรุกโลกและมีการแข่งขันระดับนานาชาติ',
+  'about.what4': 'เราเชื่อว่าหมากรุกไทยสมควรได้รับการยอมรับระดับโลกเทียบเท่าหมากรุกสากล ช่วยเราเผยแพร่!',
+  'about.diff_eyebrow': 'อะไรที่ต่างออกไป',
+  'about.diff_title': 'Makruk ให้ความรู้สึกต่างจากหมากรุกสากลอย่างไร',
+  'about.diff1': 'เม็ดอ่อนกว่าควีนสากลมาก เกมรุกจึงค่อยเป็นค่อยไปและการวางตำแหน่งสำคัญขึ้น',
+  'about.diff2': 'รูปแบบการเดินของโคนเปลี่ยนแนวทแยงและทำให้เกิดแทคติกที่ผู้เล่นหมากรุกสากลต้องเรียนรู้ใหม่',
+  'about.diff3': 'เบี้ยเริ่มบนแถวที่สามและหงายไม่เหมือนเดิม ทำให้โอเพนนิงและเอนด์เกมมีเอกลักษณ์ของตัวเอง',
+  'about.diff4': 'กฎการนับทำให้เอนด์เกมบางแบบมีจังหวะต่างจากหมากรุกสากลและต้องเล่นอย่างแม่นยำ',
+  'about.pieces_eyebrow': 'พื้นฐานตัวหมาก',
+  'about.pieces_title': 'จำการเดินของตัวหมากให้เร็ว',
+  'about.learn_by_playing': 'เรียนรู้ต่อด้วยเกมเร็ว',
+  'about.opensource_title': 'โอเพนซอร์ส',
+  'about.opensource_desc': 'หมากรุกไทยเป็นโอเพนซอร์ส 100% ภายใต้ MIT license โค้ดทั้งหมดอยู่บน GitHub',
+  'about.footer': 'หมากรุกไทย — ฟรีและโอเพนซอร์ส — สร้างด้วยรักเพื่อหมากรุกไทย',
+
+  // Games Page
+  'games.title': 'เกมล่าสุด',
+  'games.count': 'เล่นแล้ว {count} เกม',
+  'games.empty': 'ยังไม่มีเกม',
+  'games.empty_desc': 'เป็นคนแรกที่เล่น!',
+  'games.start': 'เริ่มเล่น',
+  'games.featured_title': 'เกมที่น่าดูก่อน',
+  'games.featured_desc': 'เกมที่ยาวขึ้นหรือเล่นกันจริงจังมากขึ้นจะถูกดันขึ้นมาก่อน เพื่อให้คุณเปิดวิเคราะห์เกมที่มีสาระได้เร็วขึ้น',
+  'games.no_featured_title': 'หน้านี้มีแต่เกมจบเร็วเป็นส่วนใหญ่',
+  'games.no_featured_desc': 'ผลล่าสุดชุดนี้ส่วนใหญ่จบจากการตกลงเสมอ ยอมแพ้ทันที หรือหมดเวลาเร็วเกินไป',
+  'games.low_signal_title': 'เกมจบเร็วและเกมสัญญาณต่ำ',
+  'games.low_signal_desc': 'เกมเหล่านี้ยังนับอยู่ แต่โดยมากให้บทเรียนได้น้อยกว่า เพราะจบเร็วเกินไป',
+  'games.low_signal_badge': 'จบเร็ว',
+  'games.col_game': 'เกม',
+  'games.col_time': 'เวลา',
+  'games.col_result': 'ผลลัพธ์',
+  'games.col_moves': 'ตาเดิน',
+  'games.col_when': 'เมื่อไหร่',
+  'games.prev': 'ก่อนหน้า',
+  'games.next': 'ถัดไป',
+  'games.page': 'หน้า {current} จาก {total}',
+  'games.view_leaderboard': 'ตารางคะแนน',
+  'games.filter_all': 'ทั้งหมด',
+  'games.filter_rated': 'จัดอันดับ',
+  'games.filter_casual': 'ทั่วไป',
+  'games.filter_bot': 'บอท',
+  'games.bot_badge': 'บอท',
+  'games.bot_games': 'เกมกับบอท',
+  'games.bot_win_rate': 'อัตราชนะกับบอท',
+  'games.bot_highest_level': 'ระดับสูงสุดที่ชนะได้',
+  'games.reason_checkmate': 'รุกจน',
+  'games.reason_resignation': 'ยอมแพ้',
+  'games.reason_timeout': 'หมดเวลา',
+  'games.reason_stalemate': 'อับ',
+  'games.reason_agreement': 'ตกลง',
+  'games.reason_counting': 'กฎการนับ',
+  'games.reason_draw': 'เสมอ',
+  'live.title': 'ชมเกมสด',
+  'live.desc': 'ค้นหาเกมหมากรุกไทยสาธารณะที่กำลังเล่น และตามดูทุกตาเดินในโหมดผู้ชม',
+  'live.live_now': 'กำลังเล่น',
+  'live.live_count': 'เกมสด {count} เกม',
+  'live.recently_finished': 'เพิ่งจบไป',
+  'live.recently_finished_desc': 'เกมสาธารณะเหล่านี้เพิ่งจบ และยังเปิดทบทวนได้',
+  'live.empty_title': 'ตอนนี้ยังไม่มีเกมสด',
+  'live.empty_desc': 'เมื่อมีเกม quick play สาธารณะเริ่มเล่น เกมเหล่านั้นจะปรากฏที่นี่อัตโนมัติ',
+  'live.card_live': 'สด',
+  'live.card_finished': 'จบแล้ว',
+  'live.watchers': 'ผู้ชม {count} คน',
+  'live.watch': 'ชมเกม',
+  'live.view_spectator': 'เปิดโหมดผู้ชม',
+  'live.mode_quick_play': 'Quick Play',
+  'live.mode_rated': 'Quick Play จัดอันดับ',
+  'live.mode_casual': 'Quick Play ทั่วไป',
+  'live.moves': '{count} ตาเดิน',
+  'live.updated': 'ตาเดินล่าสุดเมื่อ {n} นาทีที่แล้ว',
+  'live.just_now': 'อัปเดตเมื่อสักครู่',
+
+  // Feedback
+  'feedback.button': 'แจ้งปัญหา',
+  'feedback.title': 'ส่งความคิดเห็น',
+  'feedback.bug': 'บั๊ก',
+  'feedback.feature': 'ฟีเจอร์',
+  'feedback.other': 'อื่นๆ',
+  'feedback.placeholder_bug': 'เกิดอะไรขึ้น? คุณคาดหวังอะไร?',
+  'feedback.placeholder_feature': 'คุณอยากได้ฟีเจอร์อะไร?',
+  'feedback.placeholder_other': 'บอกเราได้เลย...',
+  'feedback.thanks': 'ขอบคุณ!',
+  'feedback.thanks_desc': 'ความคิดเห็นของคุณช่วยพัฒนา ThaiChess ออนไลน์',
+  'feedback.error': 'ส่งไม่สำเร็จ กรุณาลองใหม่',
+
+  // Feedback Messages Page
+  'feedback_page.title': 'ข้อความแจ้งปัญหา',
+  'feedback_page.count': '{count} ข้อความ',
+  'feedback_page.filter_all': 'ทั้งหมด',
+  'feedback_page.empty': 'ยังไม่มีข้อความ',
+  'feedback_page.empty_desc': 'ข้อความจากผู้ใช้จะแสดงที่นี่',
+  'feedback_page.detail_page': 'หน้า',
+  'feedback_page.detail_date': 'วันที่',
+  'fair_play.report_action': 'รายงานการเล่นต้องสงสัย',
+  'fair_play.report_sent': 'ส่งรายงานแฟร์เพลย์เพื่อตรวจสอบแล้ว',
+  'fair_play.report_sent_short': 'ส่งรายงานแล้ว',
+  'fair_play.report_failed': 'ส่งรายงานแฟร์เพลย์ไม่สำเร็จ',
+  'fair_play.admin_title': 'เคสแฟร์เพลย์',
+  'fair_play.admin_desc': 'ตรวจสอบรายงานเกมจัดอันดับที่น่าสงสัย และจำกัดเฉพาะหลังแอดมินยืนยันแล้วเท่านั้น',
+  'fair_play.admin_count': '{count} เคส',
+  'fair_play.admin_empty': 'ไม่มีเคสแฟร์เพลย์ตามตัวกรองนี้',
+  'fair_play.admin_load_failed': 'โหลดเคสแฟร์เพลย์ไม่สำเร็จ',
+  'fair_play.admin_action_failed': 'อัปเดตเคสแฟร์เพลย์ไม่สำเร็จ',
+  'fair_play.admin_note_default': 'ตรวจสอบโดยแอดมิน',
+  'fair_play.filter_all': 'ทั้งหมด',
+  'fair_play.filter_open': 'เปิดอยู่',
+  'fair_play.filter_reviewed': 'ตรวจแล้ว',
+  'fair_play.filter_restricted': 'จำกัดแล้ว',
+  'fair_play.filter_dismissed': 'ยกเลิกแล้ว',
+  'fair_play.action_restrict': 'ปิดเกมจัดอันดับ',
+  'fair_play.action_clear': 'ล้างการจำกัด',
+  'fair_play.action_dismiss': 'ยกเลิกเคส',
+  'fair_play.event_count': '{count} สัญญาณ',
+  'fair_play.updated_at': 'อัปเดต',
+  'fair_play.case_id': 'เคส',
+  'fair_play.restricted_since': 'จำกัดตั้งแต่',
+
+  // Time ago
+  'time.just_now': 'เมื่อสักครู่',
+  'time.min_ago': '{n} นาทีที่แล้ว',
+  'time.hour_ago': '{n} ชั่วโมงที่แล้ว',
+  'time.day_ago': '{n} วันที่แล้ว',
+
+  // Analysis
+  'analysis.title': 'วิเคราะห์แบบเร็ว',
+  'analysis.loading': 'กำลังโหลดเกม...',
+  'analysis.analyzing': 'กำลังตรวจเกม...',
+  'analysis.analyzing_note': 'กำลังให้ Fairy-Stockfish ตรวจเกมนี้อยู่ เมื่อเอนจินประมวลผลเสร็จ รายงานจะขึ้นที่นี่',
+  'analysis.progress': 'ตาที่ {current} จาก {total}',
+  'analysis.elapsed': 'ใช้เวลาแล้ว {seconds} วินาที',
+  'analysis.slow_hint': 'เกมยาวอาจใช้เวลานานขึ้นบนเซิร์ฟเวอร์ปัจจุบัน รายงานจะอัปเดตต่อเมื่อวิเคราะห์ตำแหน่งถัดไปเสร็จ',
+  'analysis.engine_title': 'แหล่งที่มาของรีวิว',
+  'analysis.engine_note': 'รีวิวนี้ประมวลผลด้วย {engine}',
+  'analysis.provisional_title': 'รีวิวชั่วคราว',
+  'analysis.provisional_note': 'รีวิวนี้ต้องใช้การวิเคราะห์สำรองบางตำแหน่ง ({engine}) ทำให้ค่าความแม่นยำอาจยังสูงเกินจริง',
+  'analysis.accuracy': 'ความแม่นยำโดยประมาณ',
+  'analysis.eval_graph': 'ค่าประเมิน',
+  'analysis.show_best': 'แสดงตาแนะนำ',
+  'analysis.keyboard_hint': 'ใช้ลูกศรเพื่อดูตาเดิน',
+  'analysis.eval_before': 'ก่อน',
+  'analysis.eval_after': 'หลัง',
+  'analysis.best_was': 'ตาแนะนำคือ',
+  'analysis.best': 'ดีที่สุด',
+  'analysis.excellent': 'ยอดเยี่ยม',
+  'analysis.good': 'ดี',
+  'analysis.inaccuracy': 'ไม่แม่นยำ',
+  'analysis.mistake': 'ผิดพลาด',
+  'analysis.blunder': 'พลาดร้ายแรง',
+  'analysis.analyze': 'วิเคราะห์เกม',
+  'analysis.view': 'ดูการวิเคราะห์',
+  'analysis.next': 'ถัดไป',
+  'analysis.prev': 'ย้อนกลับ',
+  'analysis.note': 'รีวิวนี้ใช้เอนจินประเมินคุณภาพตาจากโอกาสชนะที่เปลี่ยนไป คล้ายแนวทางการคำนวณ accuracy ของ Lichess',
+  'analysis.move_accuracy': 'ความแม่นยำของตานี้',
+  'analysis.desc_best': 'ตานี้ตรงกับแนวเดินที่เอนจินมองว่าแข็งแรงที่สุด',
+  'analysis.desc_good': 'เป็นตาที่ดีและยังรักษาโอกาสชนะไว้ได้เกือบทั้งหมด',
+  'analysis.desc_inaccuracy': 'ตานี้ทำให้โอกาสชนะลดลงบ้าง แต่รูปยังเล่นต่อได้',
+  'analysis.desc_mistake': 'ตานี้ทำให้โอกาสชนะลดลงพอสมควร และเอนจินพบแนวที่ดีกว่า',
+  'analysis.desc_blunder': 'ตานี้ทำให้โอกาสชนะหายไปมาก และเอนจินพบแนวที่ดีกว่าชัดเจน',
+
+  // Auth / Account
+  'auth.sign_in': 'เข้าสู่ระบบ',
+  'auth.sign_in_desc': 'บัญชีผู้ใช้เป็นทางเลือก คุณยังเล่นแบบไม่ระบุตัวตนได้',
+  'auth.hero_title': 'เก็บเรตติ้งและโปรไฟล์ของคุณไว้กับบัญชีเดียว',
+  'auth.hero_desc': 'ใช้รหัสครั้งเดียวที่ส่งไปทางอีเมล คุณยังเล่นแบบสบาย ๆ โดยไม่ต้องมีบัญชีได้ แต่การเข้าสู่ระบบจะช่วยเก็บเรตติ้ง ชื่อผู้ใช้ และประวัติเกมไว้กับคุณ',
+  'auth.benefit_optional': 'ยังเล่นแบบไม่ระบุตัวตนได้สำหรับเกมสบาย ๆ',
+  'auth.benefit_rated': 'เกมจัดอันดับและการตั้งค่าบัญชีจะผูกกับอีเมลของคุณ',
+  'auth.benefit_fast': 'ไม่ต้องจำรหัสผ่าน แค่อีเมล รหัส แล้วเข้าเล่นได้เลย',
+  'auth.back_to_play': 'กลับไปเล่น',
+  'auth.step_email': 'อีเมล',
+  'auth.step_code': 'รหัส',
+  'auth.email': 'อีเมล',
+  'auth.email_hint': 'เราจะส่งรหัส 6 หลักแบบใช้ครั้งเดียวไปที่อีเมลนี้',
+  'auth.code': 'รหัส 6 หลัก',
+  'auth.code_hint': 'โดยปกติรหัสจะมาถึงภายใน 1 นาที ถ้ายังไม่เห็นให้ตรวจสอบโฟลเดอร์สแปม',
+  'auth.send_code': 'ส่งรหัสเข้าสู่ระบบ',
+  'auth.sending_code': 'กำลังส่งรหัส...',
+  'auth.code_sent_notice': 'ตรวจสอบอีเมลของคุณเพื่อรับรหัสเข้าสู่ระบบ 6 หลัก',
+  'auth.code_sent_to': 'ส่งรหัสไปที่ {email}',
+  'auth.verify_code': 'ยืนยันรหัส',
+  'auth.signing_in': 'กำลังเข้าสู่ระบบ...',
+  'auth.use_another_email': 'ใช้อีเมลอื่น',
+  'auth.send_code_failed': 'ส่งรหัสไม่สำเร็จ',
+  'auth.sign_in_failed': 'เข้าสู่ระบบไม่สำเร็จ',
+  'account.title': 'บัญชี',
+  'account.username': 'ชื่อผู้ใช้',
+  'account.username_placeholder': '3-20 ตัวอักษร ตัวเลข หรือขีดล่าง',
+  'account.rating': 'เรตติ้ง',
+  'account.rated_games': 'เกมจัดอันดับ',
+  'account.wins': 'ชนะ',
+  'account.losses': 'แพ้',
+  'account.draws': 'เสมอ',
+  'account.saving': 'กำลังบันทึก...',
+  'account.save_profile': 'บันทึกโปรไฟล์',
+  'account.profile_updated': 'อัปเดตโปรไฟล์แล้ว',
+  'account.update_failed': 'อัปเดตโปรไฟล์ไม่สำเร็จ',
+  'account.open_feedback': 'เปิดหน้าจัดการความคิดเห็น',
+  'account.open_fair_play': 'เปิดหน้าจัดการแฟร์เพลย์',
+  'account.rated_restricted_title': 'จำกัดเกมจัดอันดับ',
+  'account.rated_restricted_desc': 'บัญชีนี้ยังเข้าสู่ระบบและเล่นเกมทั่วไปได้ แต่ปิดการจับคู่แบบจัดอันดับไว้ระหว่างรอตรวจสอบแฟร์เพลย์',
+  'account.sign_out': 'ออกจากระบบ',
+  'account.puzzle_eyebrow': 'ความคืบหน้าปริศนา',
+  'account.puzzle_title': 'ภาพรวมการฝึก',
+  'account.puzzle_desc': 'ประวัติปริศนาของคุณจะติดไปกับบัญชีและกลับมาได้ทุกอุปกรณ์',
+  'account.puzzle_completed_label': 'แก้แล้ว',
+  'account.puzzle_remaining_label': 'เหลือ',
+  'account.puzzle_focus_label': 'ธีมเด่น',
+  'account.puzzle_focus_empty': 'ยังไม่มีธีม',
+  'account.puzzle_next_label': 'เล่นต่อที่',
+  'account.puzzle_continue': 'ฝึกต่อ',
+  'account.puzzle_all_done': 'คุณเคลียร์ชุดปัจจุบันแล้ว',
+  'account.puzzle_all_done_desc': 'เปิดดูรายการทั้งหมดอีกครั้งและทบทวนโจทย์ที่ยากที่สุด',
+  'account.puzzle_last_played_label': 'เล่นล่าสุด',
+  'account.puzzle_last_played_empty': 'เริ่มเล่นปริศนาแล้วมันจะมาแสดงตรงนี้',
+  'account.puzzle_last_played_meta': '{status} · {date}',
+  'account.puzzle_recent_label': 'ปริศนาที่เพิ่งแก้',
+  'account.puzzle_recent_empty': 'ยังไม่มีปริศนาที่แก้สำเร็จ',
+  'account.puzzle_recent_meta': 'แก้เมื่อ {date}',
+  'account.puzzle_status_in_progress': 'กำลังเล่น',
+  'account.puzzle_status_solved': 'แก้แล้ว',
+  'account.puzzle_percent': 'สำเร็จแล้ว {percent}% ของชุดปัจจุบัน {total} ปริศนา',
+
+  // Leaderboard
+  'leaderboard.eyebrow': 'อันดับเกมจัดอันดับ',
+  'leaderboard.title': 'ตารางคะแนน',
+  'leaderboard.desc': 'ผู้เล่นเรตติ้งสูงสุดจากเกมควิกเพลย์แบบจัดอันดับที่จบแล้ว',
+  'leaderboard.count': 'ผู้เล่นติดอันดับ {count} คน',
+  'leaderboard.empty': 'ยังไม่มีผู้เล่นจัดอันดับ',
+  'leaderboard.empty_desc': 'เล่นควิกเพลย์แบบจัดอันดับให้จบสักเกม แล้วชื่อจะขึ้นที่นี่',
+  'leaderboard.col_rank': 'อันดับ',
+  'leaderboard.col_player': 'ผู้เล่น',
+  'leaderboard.col_rating': 'เรตติ้ง',
+  'leaderboard.col_games': 'เกม',
+  'leaderboard.col_record': 'สถิติ',
+  'leaderboard.col_winrate': 'อัตราชนะ',
+  'leaderboard.you': 'คุณ',
+  'leaderboard.view_recent': 'เกมล่าสุด',
+  'leaderboard.mobile_stats': '{games} เกม • {record} • {winrate}',
+
+  // Error UI
+  'error.board_display': 'เกิดข้อผิดพลาดในการแสดงกระดาน',
+  'error.connection_title': 'ข้อผิดพลาดการเชื่อมต่อ',
+  'error.connection_body': 'เกิดข้อผิดพลาดขณะโหลดข้อมูล',
+  'error.try_again': 'ลองอีกครั้ง',
+  'error.something_wrong': 'เกิดข้อผิดพลาดบางอย่าง',
+  'error.unexpected': 'แอปพบข้อผิดพลาดที่ไม่คาดคิด และได้บันทึกไว้แล้ว',
+  'error.unknown': 'ข้อผิดพลาดที่ไม่ทราบสาเหตุ',
+  'error.reload_page': 'โหลดหน้าใหม่',
+  'error.report_bug': 'รายงานบั๊กนี้บน GitHub',
+
+  // Header
+  'header.admin': 'แอดมิน',
+  'header.sign_in': 'เข้าสู่ระบบ',
+  'header.menu': 'เมนู',
+  'header.close_menu': 'ปิด',
+
+  // Language
+  'lang.switch': 'EN',
+};
+const loadedTranslations: Partial<Record<Language, TranslationCatalog>> = {
+  en: EN_TRANSLATIONS,
+};
+
+export async function ensureTranslations(lang: Language): Promise<TranslationCatalog> {
+  if (loadedTranslations[lang]) {
+    return loadedTranslations[lang];
+  }
+
+  if (lang === 'th') {
+    const thaiTranslations = await loadThaiTranslations();
+    loadedTranslations.th = thaiTranslations;
+    return thaiTranslations;
+  }
+
+  return EN_TRANSLATIONS;
+}
+
+export async function preloadDetectedTranslations(): Promise<Language> {
+  const lang = detectLanguage();
+  await ensureTranslations(lang);
+  return lang;
+}
+
+function getTranslations(lang: Language): TranslationCatalog {
+  return loadedTranslations[lang] ?? EN_TRANSLATIONS;
+}
+
+export function translate(
+  key: string,
+  params?: Record<string, string | number>,
+  lang: Language = detectLanguage()
+): string {
+  const translations = getTranslations(lang);
+  return applyParams(translations[key] || EN_TRANSLATIONS[key] || key, params);
+}
+
+export function I18nProvider({ children }: { children: ReactNode }) {
+  const [lang, setLangState] = useState<Language>(detectLanguage);
+  const [, setCatalogVersion] = useState(0);
+
+  const setLang = useCallback((newLang: Language) => {
+    const applyLanguage = () => {
+      setLangState(newLang);
+      localStorage.setItem('thaichess-lang', newLang);
+      document.documentElement.lang = newLang;
+    };
+
+    if (newLang === 'en' || loadedTranslations[newLang]) {
+      applyLanguage();
+      return;
+    }
+
+    void ensureTranslations(newLang).then(() => {
+      applyLanguage();
+    });
+  }, []);
+
+  useEffect(() => {
+    document.documentElement.lang = lang;
+  }, [lang]);
+
+  useEffect(() => {
+    if (lang === 'en' || loadedTranslations[lang]) {
+      return;
+    }
+
+    let cancelled = false;
+
+    void ensureTranslations(lang).then(() => {
+      if (cancelled) return;
+      setCatalogVersion((version) => version + 1);
+    });
+
+    return () => {
+      cancelled = true;
+    };
+  }, [lang]);
+
+  const t = useCallback((key: string, params?: Record<string, string | number>): string => {
+    const translations = getTranslations(lang);
+    return applyParams(translations[key] || EN_TRANSLATIONS[key] || key, params);
+  }, [lang]);
+
+  return (
+    <I18nContext.Provider value={{ lang, setLang, t }}>
+      {children}
+    </I18nContext.Provider>
+  );
+}
+
+export function useTranslation() {
+  const ctx = useContext(I18nContext);
+  if (!ctx) throw new Error('useTranslation must be used within I18nProvider');
+  return ctx;
+}
